Analysis of periodic systems

pspole Computation of the poles of a periodic system.

pszero Computation of the zeros of a periodic system.

isstable Assessment of stability of a periodic system.

PeriodicSystems.pspole — Function

pspole(psys::PeriodicStateSpace{PM}[,K]; kwargs...) -> val

Return for the periodic system psys = (A(t),B(t),C(t),D(t)) the complex vector val containing the characteristic exponents of the periodic matrix A(t) (also called the poles of the system psys).

Depending on the underlying periodic matrix type PM, the optional argument K and keyword arguments kwargs may have the following values:

if PM = PeriodicFunctionMatrix, or PM = PeriodicSymbolicMatrix, or PM = PeriodicTimeSeriesMatrix, then K is the number of factors used to express the monodromy matrix of A(t) (default: K = 1) and kwargs are the keyword arguments of pseig(::PeriodicFunctionMatrix);

if PM = HarmonicArray, then K is the number of harmonic components used to represent the Fourier series of A(t) (default: K = max(10,nh-1), nh is the number of harmonics terms of A(t)) and kwargs are the keyword arguments of psceig(::HarmonicArray);

if PM = FourierFunctionMatrix, then K is the number of harmonic components used to represent the Fourier series of A(t) (default: K = max(10,nh-1), nh is the number of harmonics terms of A(t)) and kwargs are the keyword arguments of psceig(::FourierFunctionMatrix);

if PM = PeriodicMatrix or PM = PeriodicArray, then K is the starting sample time (default: K = 1) and kwargs are the keyword arguments of psceig(::PeriodicMatrix);

PeriodicSystems.pszero — Function

pszero(psys::PeriodicStateSpace{HarmonicArray}[, N]; P, atol, rtol, fast) -> val
pszero(psys::PeriodicStateSpace{PeriodicFunctionMatrix}[, N]; P, atol, rtol, fast) -> val
pszero(psys::PeriodicStateSpace{PeriodicSymbolicMatrix}[, N]; P, atol, rtol, fast) -> val

Compute the finite and infinite zeros of a continuous-time periodic system psys = (A(t), B(t), C(t), D(t)) in val, where the periodic system matrices A(t), B(t), C(t), and D(t) are in a Harmonic Array, or Periodic Function Matrix, or Periodic Symbolic Matrix representation (the last two representation are automatically converted to a Harmonic Array representation). N is the number of selected harmonic components in the Fourier series of the system matrices (default: N = max(20,nh-1), where nh is the maximum number of harmonics terms) and the keyword parameter P is the number of full periods to be considered (default: P = 1) to build a frequency-lifted LTI representation based on truncated block Toeplitz matrices (see ps2fls).

The computation of the zeros of the complex lifted system is performed by reducing the corresponding system pencil to an appropriate Kronecker-like form which exhibits the finite and infinite eigenvalues. The reduction is performed using orthonal similarity transformations and involves rank decisions based on rank revealing QR-decompositions with column pivoting, if fast = true, or, the more reliable, SVD-decompositions, if fast = false. For a system psys of period T, the finite zeros are determined as those eigenvalues which have imaginary parts in the interval [-ω/2, ω/2], where ω = 2π/(P*T). To eliminate possible spurious finite eigenvalues, the intersection of two finite eigenvalue sets is computed for two lifted systems obtained for N and N+2 harmonic components. The infinite zeros are determined as the infinite zeros of the LTI system (A(ti), B(ti), C(ti), D(ti)) resulting for a random time value ti. Warning: While this evaluation of the number of infinite zeros mostly provides the correct result, there is no theoretical assessment of this approach (counterexamples are welcome!).

The keyword arguments atol and rtol specify the absolute and relative tolerances for the nonzero elements of the underlying lifted system pencil, respectively. The default relative tolerance is n*ϵ, where n is the size of the smallest dimension of the pencil, and ϵ is the working machine epsilon.

pszero(psys::PeriodicStateSpece{FourierFunctionMatrix}[, N]; P, atol, rtol, fast) -> val

Compute the finite and infinite zeros of a continuous-time periodic system psys = (Af(t), Bf(t), Cf(t), Df(t)) in val, where the periodic system matrices Af(t), Bf(t), Cf(t), and Df(t) are in a Fourier Function Matrix representation. N is the number of selected harmonic components in the Fourier series of the system matrices (default: N = max(20,nh-1), where nh is the maximum number of harmonics terms) and the keyword parameter P is the number of full periods to be considered (default: P = 1) to build a frequency-lifted LTI representation based on truncated block Toeplitz matrices (see ps2frls).

The computation of the zeros of the real lifted system is performed by reducing the corresponding system pencil to an appropriate Kronecker-like form which exhibits the finite and infinite eigenvalues. The reduction is performed using orthonal similarity transformations and involves rank decisions based on rank revealing QR-decompositions with column pivoting, if fast = true, or, the more reliable, SVD-decompositions, if fast = false. For a system psys of period T, the finite zeros are determined as those eigenvalues which have imaginary parts in the interval [-ω/2, ω/2], where ω = 2π/(P*T). To eliminate possible spurious finite eigenvalues, the intersection of two finite eigenvalue sets is computed for two lifted systems obtained for N and N+2 harmonic components. The infinite zeros are determined as the infinite zeros of the LTI system (Af(ti), Bf(ti), Cf(ti), Df(ti)) resulting for a random time value ti. Warning: While this evaluation of the number of infinite zeros mostly provides the correct result, there is no theoretical assessment of this approach (counterexamples are welcome!).

The keyword arguments atol and rtol specify the absolute and relative tolerances for the nonzero elements of the underlying lifted system pencil, respectively. The default relative tolerance is n*ϵ, where n is the size of the smallest dimension of the pencil, and ϵ is the working machine epsilon.

 pszero(psys::PeriodicStateSpace{PeriodicMatrix}[, K]; atol, rtol, fast) -> val
 pszero(psys::PeriodicStateSpace{PeriodicArray}[, K]; atol, rtol, fast) -> val

Compute the finite and infinite zeros of a discrete-time periodic system psys = (A(t), B(t), C(t), D(t)) in val, where the periodic system matrices A(t), B(t), C(t), and D(t) are in either Periodic Matrix or Periodic Array representation. The optional argument K specifies a desired time to start the sequence of periodic matrices (default: K = 1).

The computation of zeros relies on the fast structure exploiting reduction of singular periodic pencils as described in [1], which separates a matrix pencil M-λN which contains the infinite and finite zeros and the left and right Kronecker indices. The reduction is performed using orthonal similarity transformations and involves rank decisions based on rank revealing QR-decompositions with column pivoting, if fast = true, or, the more reliable, SVD-decompositions, if fast = false.

The keyword arguments atol and rtol specify the absolute and relative tolerances for the nonzero elements of the periodic system matrices, respectively. The default relative tolerance is n*ϵ, where n is proportional with the largest dimension of the periodic matrices, and ϵ is the working machine epsilon.

References

[1] A. Varga and P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems & Control Letters 50:371–381, 2003.

DescriptorSystems.isstable — Function

 isstable(psys[, K = 1]; smarg = 1, fast = false, offset = sqrt(ϵ), kwargs...) -> Bool

Return true if the continuous-time periodic system psys has only stable characteristic multipliers and false otherwise.

To assess the stability, the absolute values of the characteristic multipliers (i.e., the eigenvalues of the monodromy matrix) must be less than smarg-β, where smarg is the discrete-time stability margin (default: smarg = 1) and β is an offset specified via the keyword parameter offset = β to be used to numerically assess the stability of eigenvalues. The default value used for β is sqrt(ϵ), where ϵ is the working machine precision.

The monodromy matrix is determined as a product K state transition matrices (default: K = 1) computed by integrating numerically a homogeneous linear ODE with periodic coefficients (see function monodromy for options which can be specified via the keyword arguments kwargs). If fast = false (default) then the characteristic multipliers are computed using an approach based on the periodic Schur decomposition [1], while if fast = true the structure exploiting reduction [2] of an appropriate lifted pencil is employed. This later option may occasionally lead to inaccurate results for large number of matrices.

References

[1] A. Bojanczyk, G. Golub, and P. Van Dooren, The periodic Schur decomposition. Algorithms and applications, Proc. SPIE 1996.

[2] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

isstable(psys; smarg = 1, fast = false, offset = sqrt(ϵ)) -> Bool

Return true if the discrete-time periodic system psys has only stable characteristic multipliers and false otherwise.

To assess the stability, the absolute values of the characteristic multipliers (i.e., the eigenvalues of the monodromy matrix) must be less than smarg-β, where smarg is the discrete-time stability margin (default: smarg = 1) and β is an offset specified via the keyword parameter offset = β to be used to numerically assess the stability of eigenvalues. The default value used for β is sqrt(ϵ), where ϵ is the working machine precision.

If fast = false (default) then the characteristic multipliers are computed using an approach based on the periodic Schur decomposition [1], while if fast = true the structure exploiting reduction [2] of an appropriate lifted pencil is employed. This later option may occasionally lead to inaccurate results for large number of matrices.

References

[1] A. Bojanczyk, G. Golub, and P. Van Dooren, The periodic Schur decomposition. Algorithms and applications, Proc. SPIE 1996.

[2] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.