Periodic Schur decompositions

phess Periodic Hessenberg decomposition of a product of matrices.

phess! Periodic Hessenberg decomposition of a product of matrices (in place computation).

pschur Periodic Schur decompositions of products or quotient products of matrices.

pschur! Periodic Schur decompositions of products of matrices (in place computation).

psordschur! Reordering of periodic Schur decompositions of products or quotient products of matrices.

psordschur1! Reordering of periodic Schur decompositions of products or quotient products of square matrices.

pgschur Generalized real periodic Schur decomposition of a formal product of matrices.

pgschur! Generalized real periodic Schur decompositions of formal products of matrices (in place computation).

pgordschur! Reordering of generalized real periodic Schur decompositions a formal products of matrices.

PeriodicSystems.phess — Function

 phess(A; hind = 1, rev = true, withZ = true) -> (H, Z, ihess)
 phess1(A; hind = 1, rev = true, withZ = true) -> (H, Z, ihess)

Compute the Hessenberg decomposition of a product of square matrices A(p)*...*A(2)*A(1), if rev = true (default) or A(1)*A(2)*...*A(p) if rev = false, without evaluating the product. The matrices A(1), ..., A(p) are contained in the n×n×p array A such that the i-th matrix A(i) is contained in A[:,:,i]. The resulting n×n×p arrays H and Z contain the matrices H(1), ..., H(p) and the orthogonal matrices Z(1), ..., Z(p), respectively, such that for rev = true

       Z(2)' * A(1) * Z(1) = H(1),
       Z(3)' * A(2) * Z(2) = H(2),
              ...
       Z(1)' * A(p) * Z(p) = H(p),

and for rev = false

       Z(1)' * A(1) * Z(2) = H(1),
       Z(2)' * A(2) * Z(3) = H(2),
              ...
       Z(p)' * A(p) * Z(1) = H(p).

If hind = ihess, with 1 ≤ ihess ≤ p (default ihess = 1), then H(i), i = 1, ..., p are in a periodic Hessenberg form, with H(ihess) in upper Hessenberg form and H(i) upper triangular for i $\neq$ ihess. H(i) and Z(i) are contained in H[:,:,i] and Z[:,:,i], respectively. The performed orthogonal transformations are not accumulated if withZ = false, in which case Z = nothing.

The function phess is based on a wrapper for the SLICOT subroutine MB03VW (see PeriodicSystems.SLICOTtools.mb03vw!).

The function phess1 is based on wrappers for the SLICOT subroutines MB03VD (see PeriodicSystems.SLICOTtools.mb03vd!) and MB03VY (see PeriodicSystems.SLICOTtools.mb03vy!).

Missing docstring.

Missing docstring for phess!. Check Documenter's build log for details.

PeriodicSystems.pschur — Function

 pschur(A; rev = true, withZ = true) -> (S, Z, ev, ischur, α, γ)
 pschur1(A; rev = true, withZ = true) -> (S, Z, ev, ischur, α, γ)

Compute the extended periodic Schur decomposition of a square product of matrices A(p)*...*A(2)*A(1), if rev = true (default) or A(1)*A(2)*...*A(p) if rev = false, without evaluating the product. The matrices A(1), ..., A(p) are contained in the p-vector of matrices A such that the i-th matrix A(i), of dimensions m(i)×n(i), is contained in A[i]. The resulting p-vectors S and Z contain the matrices S(1), ..., S(p) and the orthogonal matrices Z(1), ..., Z(p), respectively, such that for rev = true

       Z(2)' * A(1) * Z(1) = S(1),
       Z(3)' * A(2) * Z(2) = S(2),
              ...
       Z(1)' * A(p) * Z(p) = S(p),

and for rev = false

       Z(1)' * A(1) * Z(2) = S(1),
       Z(2)' * A(2) * Z(3) = S(2),
              ...
       Z(p)' * A(p) * Z(1) = S(p).

The resulting index ischur is determined such that m(ischur) ≤ m(i), ∀i. The resulting S(i), for i = 1, ..., p are in an extended periodic Schur form, with S(ischur) in a quasi-upper trapezoidal form and S(i) upper trapezoidal for i $\neq$ ischur. S(i) and Z(i) are contained in S[i] and Z[i], respectively. The first nmin components of ev := α .* γ contain the core eigenvalues of the appropriate matrix product, where nmin = m(ischur), while the last nmax-nmin components of ev are zero, where nmax is the largest row or column dimension of A(i), for i = 1, ..., p. The eigenvalues can be alternatively expressed as α .* γ, where γ contains suitable scaling parameters to avoid overflows or underflows in the expressions of the eigenvalues. The performed orthogonal transformations are not accumulated if withZ = false, in which case Z = nothing.

The function pschur is based on wrappers for the SLICOT subroutines MB03VW (see PeriodicSystems.SLICOTtools.mb03vw!) and MB03BD (see PeriodicSystems.SLICOTtools.mb03bd!), based on algorithms proposed in [1] and [2].

The function pschur1 is based on wrappers for the SLICOT subroutines MB03VD (see PeriodicSystems.SLICOTtools.mb03vd!), MB03VY (see PeriodicSystems.SLICOTtools.mb03vy!), and MB03BD (see PeriodicSystems.SLICOTtools.mb03bd!), based on algorithms proposed in [1] and [2].

REFERENCES

[1] Bojanczyk, A., Golub, G. H. and Van Dooren, P. The periodic Schur decomposition: algorithms and applications. In F.T. Luk (editor), Advanced Signal Processing Algorithms, Architectures, and Implementations III, Proc. SPIE Conference, vol. 1770, pp. 31-42, 1992.

[2] Kressner, D. An efficient and reliable implementation of the periodic QZ algorithm. IFAC Workshop on Periodic Control Systems (PSYCO 2001), Como (Italy), August 27-28 2001. Periodic Control Systems 2001 (IFAC Proceedings Volumes), Pergamon.

 pschur(A; rev = true, withZ = true) -> (S, Z, ev, ischur, α, γ)
 pschur1(A; rev = true, withZ = true) -> (S, Z, ev, ischur, α, γ)
 pschur2(A; rev = true, withZ = true) -> (S, Z, ev, ischur, α, γ)

Compute the Schur decomposition of a product of square matrices A(p)*...*A(2)*A(1), if rev = true (default) or A(1)*A(2)*...*A(p) if rev = false, without evaluating the product. The matrices A(1), ..., A(p) are contained in the n×n×p array A such that the i-th matrix A(i) is contained in A[:,:,i]. The resulting n×n×p arrays S and Z contain the matrices S(1), ..., S(p) and the orthogonal matrices Z(1), ..., Z(p), respectively, such that for rev = true

       Z(2)' * A(1) * Z(1) = S(1),
       Z(3)' * A(2) * Z(2) = S(2),
              ...
       Z(1)' * A(p) * Z(p) = S(p),

and for rev = false

       Z(1)' * A(1) * Z(2) = S(1),
       Z(2)' * A(2) * Z(3) = S(2),
              ...
       Z(p)' * A(p) * Z(1) = S(p).

If sind = ischur, with 1 ≤ ischur ≤ p (default ischur = 1), then S(i), for i = 1, ..., p are in a periodic Schur form, with S(ischur) in quasi-upper triangular (or Schur) form and S(i) upper triangular for i $\neq$ ischur. S(i) and Z(i) are contained in S[:,:,i] and Z[:,:,i], respectively. The vector ev contains the eigenvalues of the appropriate matrix product. The eigenvalues can be alternatively expressed as α .* γ, where γ contains suitable scaling parameters to avoid overflows or underflows in the expressions of the eigenvalues. The performed orthogonal transformations are not accumulated if withZ = false, in which case Z = nothing.

The function pschur is based on wrappers for the SLICOT subroutines MB03VW (see PeriodicSystems.SLICOTtools.mb03vw!) and MB03BD (see PeriodicSystems.SLICOTtools.mb03bd!), based on algorithms proposed in [1] and [2].

The function pschur1 is based on wrappers for the SLICOT subroutines MB03VD (see PeriodicSystems.SLICOTtools.mb03vd!), MB03VY (see PeriodicSystems.SLICOTtools.mb03vy!) and MB03BD (see PeriodicSystems.SLICOTtools.mb03bd!), based on algorithms proposed in [1] and [2].

The function pschur2 is based on wrappers for the SLICOT subroutines MB03VD (see PeriodicSystems.SLICOTtools.mb03vd!), MB03VY (see PeriodicSystems.SLICOTtools.mb03vy!) and MB03VW (see PeriodicSystems.SLICOTtools.mb03vw!), based on the algorithm proposed in [1]. Known issue: MB03VW may fails for larger periods.

REFERENCES

[1] Bojanczyk, A., Golub, G. H. and Van Dooren, P. The periodic Schur decomposition: algorithms and applications. In F.T. Luk (editor), Advanced Signal Processing Algorithms, Architectures, and Implementations III, Proc. SPIE Conference, vol. 1770, pp. 31-42, 1992.

[2] Kressner, D. An efficient and reliable implementation of the periodic QZ algorithm. IFAC Workshop on Periodic Control Systems (PSYCO 2001), Como (Italy), August 27-28 2001. Periodic Control Systems 2001 (IFAC Proceedings Volumes), Pergamon.

Missing docstring.

Missing docstring for pschur!. Check Documenter's build log for details.

PeriodicSystems.psordschur! — Function

 psordschur!(S, Z, select; rev, sind)

Reorder the eigenvalues of the product Π = S[p]*...*S[2]*S[1], if rev = true (default) or Π = S[1]*S[2]*...*S[p] if rev = false, where Π is in real Schur form, such that the selected eigenvalues in the logical array select are moved into the leading positions. The p-vectors S and Z contain the matrices S[1], ..., S[p] in a periodic Schur form, with S[sind] in real Schur form, and the corresponding orthogonal transformation matrices Z[1], ..., Z[p], respectively. S and Z are overwritten by the updated matrices. A conjugate pair of eigenvalues must be either both included or both excluded via select.

 psordschur!(S, Z, select; rev, sind)

Reorder the eigenvalues of the product Π = S[:,:,p]*...*S[:,:,2]*S[:,:,1], if rev = true (default) or Π = S[:,:,1]*S[:,:,2]*...*S[:,:,p] if rev = false, where Π is in real Schur form, such that the selected eigenvalues in the logical array select are moved into the leading positions. The 3-dimensional arrays S and Z contain the matrices S[:,:,1], ..., S[:,:,p] in a periodic Schur form, with S[:,:,sind] in real Schur form, and the corresponding orthogonal transformation matrices Z[:,:,1], ..., Z[:,:,p], respectively. S and Z are overwritten by the updated matrices. A conjugate pair of eigenvalues must be either both included or both excluded via select.

PeriodicSystems.psordschur1! — Function

 psordschur1!(S, Z, select; rev, sind)

Reorder the core eigenvalues of the product Π = S[p]*...*S[2]*S[1], if rev = true (default) or Π = S[1]*S[2]*...*S[p] if rev = false, where Π is in real Schur form, such that the selected eigenvalues in the logical array select are moved into the leading positions. The p-vectors S and Z contain the matrices S[1], ..., S[p] in an extended periodic Schur form, with the leading square block of S[sind] in real Schur form, and the corresponding orthogonal transformation matrices Z[1], ..., Z[p], respectively. S and Z are overwritten by the updated matrices. A conjugate pair of eigenvalues must be either both included or both excluded via select. The dimension of select must be equal to the number of core eigenvalues (i.e., the minimum dimension of matrices in the vector S).

Missing docstring.

Missing docstring for pgschur. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pgschur!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pgordschur!. Check Documenter's build log for details.