Basic periodic system connectios and operations

psparallel Parallel connection of periodic systems.

psseries Series connection of periodic systems.

psappend Appending of periodic systems.

pshorzcat Horizontal concatenation of periodic systems.

psvertcat Vertical concatenation of periodic systems.

psinv Inversion of periodic systems.

psfeedback Periodic output feedback connections.

pssfeedback Periodic state feedback connection.

pssofeedback Periodic state feedback with state estimator connection.

PeriodicSystems.psparallel — Function

psys = psparallel(psys1, psys2)
psys = psys1 + psys2

Build the parallel connection psys of periodic systems psys1 and psys2. This coupling formally corresponds to the addition of their transfer maps as psys = psys1 + psys2.

PeriodicSystems.psseries — Function

psys = psseries(psys1, psys2)
psys = psys2*psys1

Build the series connection psys of periodic systems psys1 and psys2. This coupling formally corresponds to the product of their transfer maps as psys = psys2*psys1.

PeriodicSystems.psappend — Function

psys = psappend(psys1, psys2)

Append the periodic systems psys1 and psys2 by concatenating their input and output vectors. This corresponds to build psys as the block diagonal concatenation of their transfer maps.

PeriodicSystems.pshorzcat — Function

psys = pshorzcat(psys1,psys2)
psys = [psys1 psys2]

Concatenate horizontally the two periodic systems psys1 and psys2 by concatenating their input vectors. This formally corresponds to the horizontal concatenation of their transfer maps.

PeriodicSystems.psvertcat — Function

psys = psvertcat(psys1,psys2)
psys = [psys1; psys2]

Concatenate vertically the two periodic systems psys1 and psys2 by concatenating their output vectors. This formally corresponds to the vertical concatenation of their transfer maps.

PeriodicSystems.psinv — Function

psysi = psinv(psys)

Compute the inverse psysi of the square periodic system psys. This operation formally corresponds to the inversion of the transfer map of psys such that psysi*psys is the identity mapping.

PeriodicSystems.psfeedback — Function

 psyscl = psfeedback(psys, K, (inp, out); negative = false)

Build for a given periodic system psys with input vector u and output vector y and a periodic output feedback gain K(t) the closed-loop periodic system psyscl corresponding to the memoryless output feedback u[inp] = K(t)*y[out] + v, where inp and out are indices, vectors of indices, index ranges, : or any combinations of them. Only distinct indices can be specified. If negative = true, a negative psfeedback u[inp] = -K(t)*y[out] + v is used.

 psyscl = psfeedback(sys, K, (inp, out); negative = false)

Build for a given standard state-space system sys with input vector u and output vector y and a periodic output feedback gain K(t) the closed-loop periodic system psyscl corresponding to the memoryless output feedback u[inp] = K(t)*y[out] + v, where inp and out are are indices, vectors of indices, index ranges, : or any combinations of them. Only distinct indices can be specified. If negative = true, a negative feedback u[inp] = -K(t)*y[out] + v is used. For a continuous-time system sys, K must be a periodic switching matrix or a discrete-time periodic matrix, while for a discrete-time system sys, K must be a discrete-time periodic matrix with the same sample time.

PeriodicSystems.pssfeedback — Function

 psyscl = pssfeedback(psys, F, inp; negative = false)

Build for a given periodic system psys with input vector u and output vector y and a periodic state feedback gain F(t) the closed-loop periodic system psyscl corresponding to the memoryless state feedback u[inp] = F(t)*x + v, where inp are indices, a vector of indices, an index range, : or any combinations of them. Only distinct indices can be specified. If negative = true, a negative state feedback u[inp] = -F(t)*x + v is used.

PeriodicSystems.pssofeedback — Function

 psyscl = pssofeedback(psys, F, K, (inp, out); negative = false)

Build for a given periodic system psys, with input vector u and output vector y, a periodic state feedback gain F(t) and a periodic Kalman gain K(t) the closed-loop periodic system psyscl corresponding to the memoryless state feedback u[inp] = F(t)*xe + v and a full state estimator with state xe and inputs [u[inp]; y[out]], where inp and out are indices, vectors of indices, index ranges, : or any combinations of them. Only distinct indices can be specified. If negative = true, a negative state feedback u[inp] = -F(t)*xe + v is used.