Periodic matrix conversions

ts2hr Harmonic approximation of a periodic time series matrix.

ts2pfm Periodic function matrix of an interpolated periodic time series matrix.

hr2psm Periodic symbolic matrix of a harmonic representation.

PeriodicSystems.ts2hr — Function

 ts2hr(A::PeriodicTimeSeriesMatrix; atol = 0, rtol = √ϵ, n) -> Ahr::HarmonicArray

Compute the harmonic (Fourier) approximation of a periodic matrix specified by a time series matrix. The periodic matrix A(t) is specified as a continuous-time periodic time series matrix A, with m matrices contained in the vector of matrices A.values, where A.values[k] is the value of A(t) at time moment (k-1)T/m, with T = A.period being the period. The resulting harmonic approximation Ahr(t) of A(t) has the form

                 p
 Ahr(t) = A_0 +  ∑ ( Ac_i*cos(i*t*2*π/T)+As_i*sin(i*2*π*t/T) ) 
                i=1

where A_0 is the constant term (the mean value), Ac_i and As_i are the coefficient matrices of the i-th cosinus and sinus terms, respectively. The order of the approximation p is determined using the maximum order specified by n (default: n = (m-1)/2) and the absolute and relative tolerances atol and rtol, as follows: p is the minimum between n, (m-1)/2 and the maximum index k such that Ac_k and/or As_k are nonzero. The tolerance used to assess nonzero elements is tol = max(atol, rtol*maxnorm), where maxnorm is the maximum norm of the matrices contained in A.values. The default values of tolerances are atol = 0 and rtol = √ϵ, where ϵ is the working machine precision.

The resulting harmonic approximation Ahr(t) is returned in the harmonic array object Ahr (see HarmonicArray).

PeriodicSystems.ts2pfm — Function

 ts2pfm(At::PeriodicTimeSeriesMatrix; method = "linear") -> A::PeriodicFunctionMatrix

Compute the periodic function matrix corresponding to an interpolated periodic time series matrix. For the given periodic time series matrix At, a periodic function matrix A(t) is defined as the mapping A(t) = t -> etpf(t), where etpf(t) is a periodic interpolation/extrapolation object, as provided in the Interpolations.jl package. The keyword parameter method specifies the interpolation/extrapolation method to be used as follows:

method = "constant" - use periodic B-splines of degree 0 (periodic constant interpolation);

method = "linear" - use periodic B-splines of degree 1 (periodic linear interpolation) (default);

method = "quadratic" - use periodic B-splines of degree 2 (periodic quadratic interpolation);

method = "cubic" - use periodic B-splines of degree 3 (periodic cubic interpolation).

PeriodicSystems.hr2psm — Function

 hr2psm(Ahr::HarmonicArray, nrange) -> A::Matrix{Num}

Convert a range of harmonic components nrange of the harmonic array Ahr to a symbolic matrix A. The default range is nrange = 0:n, where n is the order of the maximum harmonics.