Periodic Lyapunov matrix equation solvers

prlyap Solution of reverse-time periodic discrete-time Lyapunov equation.

pflyap Solution of forward-time periodic discrete-time Lyapunov equations.

PeriodicSystems.prlyap — Function

prlyap(A, C) -> X

Solve the reverse-time periodic discrete-time Lyapunov equation

A'σXA + C = X

where σ is the forward shift operator σX(i) = X(i+1).

For the periodic matrices A and C with the same sampling period and commensurate periods, the periodic matrix X is determined, whose period is automatically set to the the least common period of A and C.

PeriodicSystems.pflyap — Function

pflyap(A, C) -> X

Solve the forward-time periodic discrete-time Lyapunov equation

A'XA + C = σX

where σ is the forward shift operator σX(i) = X(i+1).

For the periodic matrices A and C with the same sampling period and commensurate periods, the periodic matrix X is determined, whose period is automatically set to the the least common period of A and C.