Periodic Riccati equation solvers

pcric Solution of periodic Riccati differential equations.

prcric Solution of control-related reverse-time periodic Riccati differential equation.

pfcric Solution of filtering-related forward-time periodic Riccati differential equation.

pgcric Computation of periodic generators for periodic Riccati differential equations.

PeriodicSystems.pcric — Function

pcric(A, R, Q; K = 10, adj = false, solver, reltol, abstol, fast) -> (X, EVALS)

Solve the periodic Riccati differential equation

.                                                 
X(t) = A(t)X(t) + X(t)A(t)' + Q(t) - X(t)R(t)X(t) ,  if adj = false,

or

 .                                                
-X(t) = A(t)'X(t) + X(t)A(t) + Q(t) - X(t)R(t)X(t) , if adj = true.

The periodic matrices A, R and Q must have the same type, the same dimensions and commensurate periods, and additionally R and Q must be symmetric. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A, R and Q and the number of subperiods is adjusted accordingly.

If fast = true (default) the multiple-shooting method is used in conjunction with fast pencil reduction techniques, as proposed in [1], to determine the periodic solution in t = 0 and a multiple point generator of the appropriate periodic differential Riccati equation is determined (see [2] for details). If fast = false, the multiple-shooting method is used in conjunction with the periodic Schur decomposition to determine multiple point generators directly from the stable periodic invariant subspace of an appropriate symplectic transition matrix (see also [2] for more details).

The keyword argument K specifies the number of grid points to be used for the discretization of the continuous-time problem (default: K = 10). The multiple point generator is computed by solving the appropriate discrete-time periodic Lyapunov equation using the periodic Schur method of [2]. The resulting periodic time-series matrix is finally converted to the appropriate periodic representation.

The ODE solver to be employed to convert the continuous-time problem into a discrete-time problem can be specified using the keyword argument solver, together with the required relative accuracy reltol (default: reltol = 1.e-4) and absolute accuracy abstol (default: abstol = 1.e-7) (see tvstm). For large values of K, parallel computation of the matrices of the discrete-time problem can be alternatively performed by starting Julia with several execution threads. The number of execution threads is controlled either by using the -t/--threads command line argument or by using the JULIA_NUM_THREADS environment variable.

References

[1] A. Varga. On solving periodic differential matrix equations with applications to periodic system norms computation. Proc. IEEE CDC/ECC, Seville, 2005.

[2] A. Varga. On solving periodic Riccati equations. Numerical Linear Algebra with Applications, 15:809-835, 2008.

PeriodicSystems.prcric — Function

prcric(A, B, R, Q; K = 10, solver, reltol, abstol) -> (X, EVALS, F)

Compute the symmetric stabilizing solution X(t) of the periodic control related Riccati differential equation

 .                                                -1 
-X(t) = A(t)'X(t) + X(t)A(t) + Q(t) - X(t)B(t)R(t)  B(t)'X(t) ,

the periodic stabilizing state-feedback gain

            -1
F(t) = -R(t)  B(t)'X(t)

and the corresponding stable characteristic multipliers EVALS of A(t)-B(t)F(t).

If fast = true (default) the multiple-shooting method is used in conjunction with fast pencil reduction techniques, while if fast = false, the multiple-shooting method is used in conjunction with the periodic Schur decomposition approach.

The periodic matrices A, B, R and Q must have the same type and commensurate periods, and additionally R must be symmetric positive definite and Q must be symmetric positive semidefinite. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A, B, R and Q and the number of subperiods is adjusted accordingly.

PeriodicSystems.pfcric — Function

pfcric(A, C, R, Q; K = 10, solver, reltol, abstol, fast) -> (X, EVALS, F)

Compute the symmetric stabilizing solution X(t) of the periodic filtering related Riccati differential equation

.                                                 -1 
X(t) = A(t)X(t) + X(t)A(t)' + Q(t) - X(t)C(t)'R(t)  C(t)X(t) ,

the periodic stabilizing Kalman gain

                     -1
F(t) = -X(t)C(t)'R(t)

and the corresponding stable characteristic multipliers EVALS of A(t)-F(t)C(t).

If fast = true (default) the multiple-shooting method is used in conjunction with fast pencil reduction techniques, while if fast = false, the multiple-shooting method is used in conjunction with the periodic Schur decomposition approach.

The periodic matrices A, C, R and Q must have the same type and commensurate periods, and additionally R must be symmetric positive definite and Q must be symmetric positive semidefinite. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A, C, R and Q and the number of subperiods is adjusted accordingly.

PeriodicSystems.pgcric — Function

pgcric(A, B, R, Q[, K = 1]; adj = false, solver, reltol, abstol) -> (X, EVALS)

Compute periodic generators for the periodic Riccati differential equation in the filtering form

.                                                  
X(t) = A(t)X(t) + X(t)A(t)' + Q(t) - X(t)R(t)X(t), if adj = false,

or in the control form

 .                                              
-X(t) = A(t)'X(t) + X(t)A(t) + Q(t) - X(t)R(t)X(t) , if adj = true,

where A(t), R(t) and Q(t) are periodic matrices of commensurate periods, with A(t) square, R(t) symmetric and positive definite, and Q(t) symmetric and positive semidefinite. The resulting X is a collection of periodic generator matrices determined as a periodic time-series matrix with N components, where N = 1 if A(t), R(t) and Q(t) are constant matrices and N = K otherwise. EVALS contains the stable characteristic multipliers of the monodromy matrix of the corresponding Hamiltonian matrix (also called closed-loop characteristic multipliers). The period of X is set to the least common commensurate period of A(t), R(t) and Q(t) and the number of subperiods is adjusted accordingly. Any component matrix of X is a valid initial value to be used to generate the solution over a full period by integrating the appropriate differential equation. The multiple-shooting method proposed in [1] is employed, first, to convert the continuous-time periodic Lyapunov into a discrete-time periodic Lyapunov equation satisfied by the generator solution in K time grid points and then to compute the solution by solving an appropriate discrete-time periodic Lyapunov equation using the periodic Schur method of [2].

The ODE solver to be employed to convert the continuous-time problem into a discrete-time problem can be specified using the keyword argument solver, together with the required relative accuracy reltol (default: reltol = 1.e-4) and absolute accuracy abstol (default: abstol = 1.e-7) (see tvstm). For large values of K, parallel computation of the matrices of the discrete-time problem can be alternatively performed by starting Julia with several execution threads. The number of execution threads is controlled either by using the -t/--threads command line argument or by using the JULIA_NUM_THREADS environment variable.

References

[1] A. Varga. On solving periodic differential matrix equations with applications to periodic system norms computation. Proc. IEEE CDC/ECC, Seville, 2005.

[2] A. Varga. Periodic Lyapunov equations: some applications and new algorithms. Int. J. Control, vol, 67, pp, 69-87, 1997.