Stabilization of periodic systems

Periodic state feedback controller and estimator design

Periodic output feedback controller design

plqofc LQ-optimal stabilization of discrete-time periodic systems using periodic output feedback.

plqofc_sw LQ-optimal stabilization of discrete-time periodic systems using switching periodic output feedback.

PeriodicSystems.plqofc — Function

plqofc(psys, Q, R, S; G = I, sdeg = 1, optimizer = Optim.LBFGS(), vinit, maxiter, vtol, Jtol, gtol, show) -> (K, info)

Compute for the discrete-time periodic state-space system psys = (A(t),B(t),C(t),D(t)), the optimal periodic feedback gain K(t) in the output feedback control law u(t) = K(t)*y(t), which minimizes the expectation of the quadratic index:

 J = E{ Sum [x(t)'*Q(t)*x(t) + 2*x(t)'*S(t)*u(t) + u(t)'*R(t)*u(t)] },

where Q(t), R(t), S(t) are periodic weighting matrices. By default S(t) is missing, in which case, S(t) = 0 is assumed. The covariance of the initial state x(0) can be specified via the keyword argument G (default: G = I) and a desired stability degree of the closed-loop characteristic multipliers can be specified using the keyword argument sdeg (default: sdeg = 1).

For the determination of the optimal feedback gain an optimization-based approach is employed using tools available in the optimization package Optim.jl. By default, the gradient-based limited-memory quasi-Newton method (also known as L-BFGS) for unconstrained minimizations is employed, using the formulas for function and gradient evaluations developed in [1]. Alternative gradient-based methods can be selected using the keyword argument optimizer, such as, for example, the quasi-Newton method BFGS with optimizer = Optim.BFGS(), or for small size optimization problems, the Nelder-Mead gradient-free method with optimizer = Optim.NelderMead(). The keyword parameter maxiter = maxit can be used to specify the maximum number of iterations to be performed (default: maxit = 1000). The keyword argument vtol can be used to specify the absolute tolerance in the changes of the optimization variable v (default: vtol = 0). The keyword argument Jtol can be used to specify the relative tolerance in the changes of the optimization criterium J (default: Jtol = 0), while gtol can be used to specify the absolute tolerance in the gradient ∇J, in infinity norm (default: gtol = 1e-5). With the keyword argument show = true, a trace of the optimization algorithm's state is shown on stdout (default show = false).

A bound-constrained optimization can be performed using the keyword argument lub = (lb, ub), where lb is the lower bound and ub is the upper bound on the feedback gains. By default, lub = missing in which case the bounds lb = -Inf and ub = Inf are assumed.

An initial value for the optimization variable v can be provided using the keyword argument vinit = gains, with gains a m×p×N array, where m and p are the numbers of system inputs and outputs, respectively, and N is the number of samples (i.e., N = psys.period/psys.Ts). By default, vinit = missing, in which case a zero matrix gains = 0 is assumed.

References

[1] A. Varga and S. Pieters. Gradient-based approach to solve optimal periodic output feedback problems. Automatica, vol.34, pp. 477-481, 1998.

PeriodicSystems.plqofc_sw — Function

plqofc_sw(psys, Q, R, S; ns, vinit, kwargs...) -> (K, info)

Compute for the discrete-time periodic state-space system psys = (A(t),B(t),C(t),D(t)), the optimal switching periodic feedback gain K(t) in the output feedback control law u(t) = K(t)*y(t), which minimizes the expectation of the quadratic index:

 J = E{ Sum [x(t)'*Q(t)*x(t) + 2*x(t)'*S(t)*u(t) + u(t)'*R(t)*u(t)] },

where Q(t), R(t), S(t) are periodic weighting matrices. The switching times are specified by the N-dimensional integer vector ns and the resulting switching periodic gain K(t) is defined by N values (see SwitchingPeriodicArray). By default, ns = 1:N, where N is the maximal number of samples (i.e., N = psys.period/psys.Ts).

If an initial value for the optimization variable v is provided using the keyword argument vinit = gains, then gains must be an m×p×N array, where m and p are the numbers of system inputs and outputs. By default, vinit = missing, in which case a zero matrix gains = 0 is assumed.

The rest of keyword parameters contained in kwargs are the same as those used for the function plqofc.

The computation of gradient fully exploits the switching structure of the feedback gain, using formulas which generalize the case of constant feedback considered in [1].

References

[1] A. Varga and S. Pieters. Gradient-based approach to solve optimal periodic output feedback problems. Automatica, vol.34, pp. 477-481, 1998.