Periodic matrix data types

PeriodicMatrix Discrete-time periodic matrix representation.

PeriodicArray Discrete-time periodic array representation.

PeriodicFunctionMatrix Continuous-time periodic function matrix representation.

PeriodicSymbolicMatrix Continuous-time periodic symbolic matrix representation.

PeriodicTimeSeriesMatrix Continuous-time periodic time series matrix representation.

HarmonicArray Continuous-time harmonic array representation.

PeriodicSystems.PeriodicMatrix — Type

PeriodicMatrix(M, T) -> A::PeriodicMatrix

Discrete-time periodic matrix representation.

The discrete periodic matrix object A is built from a p-vector M of real matrices and the associated time period T. M contains the cyclic component matrices M[i], i = 1,..., p, where M[i] represents the value M(Δ(i-1)) of a time periodic matrix M(t) of period T, with Δ := T/p, the associated sampling time. It is assumed that M[k] := M[mod(k-1,p)+1] for arbitrary k. All component matrices are allowed to have arbitrary (time-varying) dimensions.

PeriodicSystems.PeriodicArray — Type

PeriodicArray(M, T) -> A::PeriodicArray

Discrete-time periodic array representation.

The discrete periodic matrix object A is built from a m×n×p real array M and the associated time period T. M contains the cyclic component matrices M[:,:,i], i = 1,..., p, where M[:,:,i] represents the value M(Δ(i-1)) of a time periodic matrix M(t) of period T, with Δ := T/p, the associated sampling time. It is assumed that M[:,:,k] := M[:,:,mod(k-1,p)+1] for arbitrary k.

PeriodicSystems.PeriodicFunctionMatrix — Type

PeriodicFunctionMatrix(f, T) -> A::PeriodicFunctionMatrix

Continuous-time periodic function matrix representation.

A = t -> f(t) is a time function defining a m×n time periodic real matrix A(t) of period T such that A(t) = A(t+T) for any time value t.

PeriodicSystems.PeriodicSymbolicMatrix — Type

PeriodicSymbolicMatrix(F, T) -> A::PeriodicSymbolicMatrix

Continuous-time periodic symbolic matrix representation.

F(t) is a symbolic periodic real matrix or vector of symbolic variable t and period T such that F(t) = F(t+T) for any real time value t.

PeriodicSystems.PeriodicTimeSeriesMatrix — Type

PeriodicTimeSeriesMatrix(A, T) -> At::PeriodicTimeSeriesMatrix

Continuous-time periodic time series matrix representation.

The continuous-time periodic time series matrix object At is built from a p-vector A of real matrices and the associated time period T. A contains the cyclic component matrices A[i], i = 1,..., p, where A[i] represents the value A(Δ(i-1)) of a time periodic matrix A(t) of period T, with Δ := T/p, the associated sampling time. It is assumed that A[k] := A[mod(k-1,p)+1] for arbitrary k. All component matrices must have the same dimensions.

PeriodicSystems.HarmonicArray — Type

 HarmonicArray(A, T) -> Ahr::HarmonicArray

Continuous-time harmonic array representation.

The harmonic array object Ahr is built for the harmonic representation of a periodic matrix A(t) of period T in the form

               p
 A(t) = A_0 +  ∑ ( Ac_i*cos(i*t*2*π/T)+As_i*sin(i*2*π*t/T) ) .
              i=1

The m×n×(p+1) complex array A contains the harmonic components as follows: A[:,:,1] contains the constant term A_0 (the mean value) and the real and imaginary parts of A[:,:,i+1] for i = 1, ..., p contain the coefficient matrices Ac_i and As_i, respectively.

PeriodicSystems.HarmonicArray — Method

 HarmonicArray(A0, Ac, As, T) -> Ahr::HarmonicArray

Construct a harmonic array representation from the harmonic components.

The harmonic array object Ahr is built for the harmonic representation of a periodic matrix A(t) of period T in the form

               p
 A(t) = A_0 +  ∑ ( Ac_i*cos(i*t*2*π/T)+As_i*sin(i*2*π*t/T) ) ,
              i=1

where the constant term A_0 is contained in the real matrix A0, and Ac and As are vectors of real matrices such that the i-th (cosinus) coefficient matrix Ac_i is contained in Ac[i] and the i-th (sinus) coefficient matrix As_i is contained in As[i]. p is the maximum of length of the vectors of matrices Ac and As. If the length of Ac or As is less than p, then zero trailing matrices are assumed in the respective matrix. All component matrices must have the same dimensions.