Periodic matrix conversions

ts2hr Convert a periodic time series matrix to a harmonic array approximation.

pfm2hr Convert a periodic function matrix to a harmonic array representation.

ts2pfm Convert an interpolated periodic time series matrix to a periodic function matrix.

hr2psm Convert a harmonic array representation to a periodic symbolic matrix.

psm2hr Convert a periodic symbolic matrix into a harmonic array representation.

pm2pa Convert a discrete-time periodic matrix object to a periodic array object.

PeriodicSystems.ts2hr — Function

 ts2hr(A::PeriodicTimeSeriesMatrix; atol = 0, rtol = √ϵ, n) -> Ahr::HarmonicArray

Compute the harmonic (Fourier) approximation of a periodic matrix specified by a time series matrix. The periodic matrix A(t) is specified as a continuous-time periodic time series matrix A, with m matrices contained in the vector of matrices A.values, where A.values[k] is the value of A(t) at time moment (k-1)T/m, with T = A.period being the period. The resulting harmonic approximation Ahr(t) of A(t) has the form

                 p
 Ahr(t) = A_0 +  ∑ ( Ac_i*cos(i*t*2*π/T)+As_i*sin(i*2*π*t/T) ) 
                i=1

where A_0 is the constant term (the mean value), Ac_i and As_i are the coefficient matrices of the i-th cosinus and sinus terms, respectively. The order of the approximation p is determined using the maximum order specified by n (default: n = (m-1)/2) and the absolute and relative tolerances atol and rtol, as follows: p is the minimum between n, (m-1)/2 and the maximum index k such that Ac_k and/or As_k are nonzero. The tolerance used to assess nonzero elements is tol = max(atol, rtol*maxnorm), where maxnorm is the maximum norm of the matrices contained in A.values. The default values of tolerances are atol = 0 and rtol = √ϵ, where ϵ is the working machine precision.

The resulting harmonic approximation Ahr(t) is returned in the harmonic array object Ahr (see HarmonicArray).

PeriodicSystems.pfm2hr — Function

 pfm2hr(A::PeriodicFunctionMatrix; nsample, NyquistFreq) -> Ahr::HarmonicArray

Convert a periodic function matrix into a harmonic array representation. If A(t) is a periodic function matrix of period T, then the harmonic array representation Ahr is determined by applying the fast Fourier transform to the sampled arrays A(iΔ), i = 0, ..., k, where Δ = T/k is the sampling period and k is the number of samples specified by the keyword argument nsample = k (default: k = 128). If the Nyquist frequency f is specified via the keyword argument NyquistFreq = f, then k is chosen k = 2*f*T to avoid signal aliasing.

PeriodicSystems.ts2pfm — Function

 ts2pfm(At::PeriodicTimeSeriesMatrix; method = "linear") -> A::PeriodicFunctionMatrix

Compute the periodic function matrix corresponding to an interpolated periodic time series matrix. For the given periodic time series matrix At, a periodic function matrix A(t) is defined as the mapping A(t) = t -> etpf(t), where etpf(t) is a periodic interpolation/extrapolation object, as provided in the Interpolations.jl package. The keyword parameter method specifies the interpolation/extrapolation method to be used as follows:

method = "constant" - use periodic B-splines of degree 0 (periodic constant interpolation);

method = "linear" - use periodic B-splines of degree 1 (periodic linear interpolation) (default);

method = "quadratic" - use periodic B-splines of degree 2 (periodic quadratic interpolation);

method = "cubic" - use periodic B-splines of degree 3 (periodic cubic interpolation).

PeriodicSystems.hr2psm — Function

 hr2psm(Ahr::HarmonicArray, nrange) -> A::Matrix{Num}

Convert a range of harmonic components nrange of the harmonic array Ahr to a symbolic matrix A. The default range is nrange = 0:n, where n is the order of the maximum harmonics.

PeriodicSystems.psm2hr — Function

 psm2hr(A::PeriodicSymbolicMatrix; nsample, NyquistFreq) -> Ahr::HarmonicArray

Convert a periodic symbolic matrix into a harmonic array representation. If A(t) is a periodic symbolic matrix of period T, then the harmonic array representation Ahr is determined by applying the fast Fourier transform to the sampled arrays A(iΔ), i = 0, ..., k, where Δ = T/k is the sampling period and k is the number of samples specified by the keyword argument nsample = k (default: k = 128). If the Nyquist frequency f is specified via the keyword argument NyquistFreq = f, then k is chosen k = 2*f*T to avoid signal aliasing.

PeriodicSystems.pm2pa — Function

 pm2pa(At::PeriodicMatrix) -> A::PeriodicArray

Convert a discrete-time periodic matrix object into a discrete-time periodic array object.

The discrete-time periodic matrix object At contains a p-vector At of real matrices At[i], i = 1,..., p, the associated time period T and the number of subperiods k. The resulting discrete-time periodic array object A of period T and number of subperiods k is built from a m×n×p real array A, such that A[:,:,i] contains At[i], for i = 1,..., p. For non-constant dimensions, the resulting m and n are the maximum row and column dimensions, respectively, and the resulting component matrices A[:,:,i] contain At[i], appropriately padded with zero entries.