Periodic system conversions

psaverage Computation of the averaged LTI system of a periodic system.

psc2d Discretization of a continuous-time periodic system.

psmrc2d Discretization/resampling of a multirate LTI system.

PeriodicSystems.psaverage — Function

psaverage(psysc) -> sys::DescriptorStateSpace

Compute for the continuous-time periodic system psysc = (A(t),B(t),C(t),D(t)) the corresponding time averaged linear time-invatiant system sys = (Am,Bm,Cm,Dm) over one period.

PeriodicSystems.psc2d — Function

 psc2d(psysc, Ts; solver, reltol, abstol, dt) -> psys::PeriodicStateSpace{PeriodicMatrix}

Compute for the continuous-time periodic system psysc = (A(t),B(t),C(t),D(t)) of period T and for a sampling time Ts, the corresponding discretized periodic system psys = (Ad,Bd,Cd,Dd) using a zero-order hold based discretization method.

The discretization is performed by determining the monodromy matrix as a product of K = T/Ts state transition matrices of the extended state-space matrix [A(t) B(t); 0 0] by integrating numerically the corresponding homogeneous linear ODE. The ODE solver to be employed can be specified using the keyword argument solver, together with the required relative accuracy reltol (default: reltol = 1.e-3), absolute accuracy abstol (default: abstol = 1.e-7) and/or the fixed step length dt (default: dt = Ts/10) (see tvstm). For large values of K, parallel computation of factors can be alternatively performed by starting Julia with several execution threads. The number of execution threads is controlled either by using the -t/--threads command line argument or by using the JULIA_NUM_THREADS environment variable.

PeriodicSystems.psmrc2d — Function

 psmrc2d(sys, Ts; ki, ko) -> psys::PeriodicStateSpace{PeriodicMatrix}

Perform the multirate discretization of a linear time-invariant system.

For a continuous-time state-space system sys = (A,B,C,D), a basic sampling time Ts, and the integer vectors ki and ko containing, respectively, the numbers of input and output sampling subperiods, the corresponding discretized periodic system psys = (Ap,Bp,Cp,Dp) of period T = n*Ts is determined, where n is the least common multiple of the integers contained in ki and ko. For a continuous-time system sys a zero-order hold based discretization method is used, such that the i-th input is constant during intervals of length ki[i]*Ts. An output hold device is used to provide constant intersample outputs, such that the i-th output is constant during intervals of length ko[i]*Ts. For a discrete-time system with a defined sample time Δ, an input and output resampling is performed using Δ as basic sample time and the value of Ts is ignored. If the system sample time is not defined, then the value of Ts is used as the basic sample time.