Discretization of continuous-time periodic systems

psc2d Discretization of a continuous-time periodic system.

PeriodicSystems.psc2d — Function

 psc2d(psysc, Ts; solver, reltol, abstol, dt) -> psys::PeriodicStateSpace{PeriodicMatrix}

Compute for the continuous-time periodic system psysc = (A(t),B(t),C(t),D(t)) of period T and for a sampling time Ts, the corresponding discretized periodic system psys = (Ad,Bd,Cd,Dd) using a zero-order hold based discretization method.

The discretization is performed by determining the monodromy matrix as a product of K = T/Ts state transition matrices of the extended state-space matrix [A(t) B(t); 0 0] by integrating numerically the corresponding homogeneous linear ODE. The ODE solver to be employed can be specified using the keyword argument solver, together with the required relative accuracy reltol (default: reltol = 1.e-3), absolute accuracy abstol (default: abstol = 1.e-7) and/or the fixed step length dt (default: dt = Ts/10) (see tvstm). For large values of K, parallel computation of factors can be alternatively performed by starting Julia with several execution threads. The number of execution threads is controlled either by using the -t/--threads command line argument or by using the JULIA_NUM_THREADS environment variable.