Periodic Lyapunov equation solvers

pclyap Solution of periodic Lyapunov differential equations.

prclyap Solution of reverse-time periodic Lyapunov differential equation equations.

pfclyap Solution of forward-time periodic Lyapunov differential equation equations.

pgclyap Computation of periodic generators for periodic Lyapunov differential equations.

pdlyap Solution of periodic discrete-time Lyapunov equations.

prdlyap Solution of reverse-time periodic discrete-time Lyapunov equations.

pfdlyap Solution of forward-time periodic discrete-time Lyapunov equations.

pdplyap Solution of positve periodic discrete-time Lyapunov equations.

prdplyap Solution of positve reverse-time periodic discrete-time Lyapunov equations.

pfdplyap Solution of positve forward-time periodic discrete-time Lyapunov equations.

PeriodicSystems.pclyap — Function

pclyap(A, C; K = 10, adj = false, solver, reltol, abstol, intpol, intpolmeth) -> X
pclyap(A, C; K = 10, adj = false, solver, reltol, abstol) -> X

Solve the periodic Lyapunov differential equation

.
X(t) = A(t)X(t) + X(t)A(t)' + C(t) , if adj = false,

or

 .
-X(t) = A(t)'X(t) + X(t)A(t) + C(t) , if adj = true.

The periodic matrices A and C must have the same type, the same dimensions and commensurate periods. Additionally C must be symmetric. The resulting symmetric periodic solution X has the type PeriodicFunctionMatrix and X(t) can be used to evaluate the value of X at time t. X has the period set to the least common commensurate period of A and C and the number of subperiods is adjusted accordingly.

The multiple-shooting method of [1] is employed to convert the (continuous-time) periodic differential Lyapunov equation into a discrete-time periodic Lyapunov equation satisfied by a multiple point generator of the solution. The keyword argument K specifies the number of grid points to be used for the discretization of the continuous-time problem (default: K = 10). If A and C are of type PeriodicSwitchingMatrix, then K specifies the number of grid points used between two consecutive switching time values (default: K = 1). The multiple point periodic generator is computed by solving the appropriate discrete-time periodic Lyapunov equation using the periodic Schur method of [2]. The resulting periodic generator is finally converted into a periodic function matrix which determines for a given t the function value X(t) by integrating the appropriate ODE from the nearest grid point value.

To speedup function evaluations, interpolation based function evaluations can be used by setting the keyword argument intpol = true (default: intpol = false). In this case the interpolation method to be used can be specified via the keyword argument intpolmeth = meth. The allowable values for meth are: "constant", "linear", "quadratic" and "cubic" (default) (see also ts2pfm). Interpolation is not possible if A and C are of type PeriodicSwitchingMatrix.

The ODE solver to be employed to convert the continuous-time problem into a discrete-time problem can be specified using the keyword argument solver, together with the required relative accuracy reltol (default: reltol = 1.e-4) and absolute accuracy abstol (default: abstol = 1.e-7) (see tvstm). Parallel computation of the matrices of the discrete-time problem can be alternatively performed by starting Julia with several execution threads. The number of execution threads is controlled either by using the -t/--threads command line argument or by using the JULIA_NUM_THREADS environment variable.

References

[1] A. Varga. On solving periodic differential matrix equations with applications to periodic system norms computation. Proc. IEEE CDC/ECC, Seville, 2005.

[2] A. Varga. Periodic Lyapunov equations: some applications and new algorithms. Int. J. Control, vol, 67, pp, 69-87, 1997.

PeriodicSystems.prclyap — Function

prclyap(A, C; K = 10, solver, reltol, abstol, intpol, intpolmeth) -> X
prclyap(A, C; K = 10, solver, reltol, abstol) -> X

Solve the periodic reverse-time Lyapunov differential equation

 .
-X(t) = A(t)'X(t) + X(t)A(t) + C(t).

The periodic matrices A and C must have the same type, the same dimensions and commensurate periods, and additionally C must be symmetric. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A and C and the number of subperiods is adjusted accordingly.

This function is merely an interface to pclyap (see this function for the description of keyword parameters).

PeriodicSystems.pfclyap — Function

pfclyap(A, C; K = 10, solver, reltol, abstol, intpol, intpolmeth) -> X
pfclyap(A, C; K = 10, solver, reltol, abstol) -> X

Solve the periodic forward-time Lyapunov differential equation

.
X(t) = A(t)X(t) + X(t)A(t)' + C(t) .

The periodic matrices A and C must have the same type, the same dimensions and commensurate periods, and additionally C must be symmetric. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A and C and the number of subperiods is adjusted accordingly.

This function is merely an interface to pclyap (see this function for the description of keyword parameters).

PeriodicSystems.pgclyap — Function

pgclyap(A, C[, K = 1]; adj = false, solver, reltol, abstol, dt) -> X

Compute periodic generators for the periodic Lyapunov differential equation

.
X(t) = A(t)X(t) + X(t)A(t)' + C(t) , if adj = false,

or

 .
-X(t) = A(t)'X(t) + X(t)A(t) + C(t) , if adj = true.

The periodic matrices A and C must have the same type, the same dimensions and commensurate periods, and additionally C must be symmetric. If A and C have the types PeriodicFunctionMatrix, HarmonicArray, FourierFunctionMatrix or PeriodicTimeSeriesMatrix, then the resulting X is a collection of periodic generator matrices determined as a periodic time-series matrix with N components, where N = 1 if A and C are constant matrices and N = K otherwise. If A and C have the type PeriodicSwitchingMatrix, then X is a collection of periodic generator matrices determined as a periodic switching matrix, whose switching times are the unique switching times contained in the union of the switching times of A and C. If K > 1, a refined grid of K equidistant values is used for each two consecutive switching times in the union. The period of X is set to the least common commensurate period of A and C and the number of subperiods is adjusted accordingly. Any component matrix of X is a valid initial value to be used to generate the solution over a full period by integrating the appropriate differential equation. The multiple-shooting method of [1] is employed, first, to convert the continuous-time periodic Lyapunov into a discrete-time periodic Lyapunov equation satisfied by the generator solution in the grid points and then to compute the solution by solving an appropriate discrete-time periodic Lyapunov equation using the periodic Schur method of [2].

The ODE solver to be employed to convert the continuous-time problem into a discrete-time problem can be specified using the keyword argument solver, together with the required relative accuracy reltol (default: reltol = 1.e-4), absolute accuracy abstol (default: abstol = 1.e-7) and stepsize dt' (default:dt = 0, only used ifsolver = "symplectic") (see [tvstm`](@ref)).

Parallel computation of the matrices of the discrete-time problem can be alternatively performed by starting Julia with several execution threads. The number of execution threads is controlled either by using the -t/--threads command line argument or by using the JULIA_NUM_THREADS environment variable.

References

[1] A. Varga. On solving periodic differential matrix equations with applications to periodic system norms computation. Proc. IEEE CDC/ECC, Seville, 2005.

[2] A. Varga. Periodic Lyapunov equations: some applications and new algorithms. Int. J. Control, vol, 67, pp, 69-87, 1997.

PeriodicSystems.pdlyap — Function

pdlyap(A, C; adj = true) -> X

Solve the periodic discrete-time Lyapunov equation

A'σXA + C = X  for adj = true

or

AXA' + C = σX  for adj = false,

where σ is the forward shift operator σX(i) = X(i+1).

The periodic matrices A and C must have the same type, the same dimensions and commensurate periods, and additionally C must be symmetric. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A and C and the number of subperiods is adjusted accordingly.

The periodic discrete analog of the Bartels-Stewart method based on the periodic Schur form of the periodic matrix A is employed [1].

Reference:

[1] A. Varga. Periodic Lyapunov equations: some applications and new algorithms. Int. J. Control, vol, 67, pp, 69-87, 1997.

PeriodicSystems.prdlyap — Function

prdlyap(A, C) -> X

Solve the reverse-time periodic discrete-time Lyapunov equation

A'σXA + C = X

where σ is the forward shift operator σX(i) = X(i+1).

The periodic matrices A and C must have the same type, the same dimensions and commensurate periods, and additionally C must be symmetric. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A and C and the number of subperiods is adjusted accordingly.

PeriodicSystems.pfdlyap — Function

pfdlyap(A, C) -> X

Solve the forward-time periodic discrete-time Lyapunov equation

AXA' + C = σX

where σ is the forward shift operator σX(i) = X(i+1).

The periodic matrices A and C must have the same type, the same dimensions and commensurate periods, and additionally C must be symmetric. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A and C and the number of subperiods is adjusted accordingly.

PeriodicSystems.pdplyap — Function

pdplyap(A, C; adj = true) -> U

Compute the upper triangular factor U of the solution X = U'U of the periodic discrete-time Lyapunov matrix equation

  A'σXA + C'C = X, if adj = true,

or of the solution X = UU' of the periodic discrete-time Lyapunov matrix equation

  AXA' + CC' =  σX, if adj = false,

where σ is the forward shift operator σX(i) = X(i+1). The periodic matrix A must be stable, i.e., have all characteristic multipliers with moduli less than one.

The periodic matrices A and C must have the same type, the same dimensions and commensurate periods. The resulting upper triangular periodic matrix U has the period set to the least common commensurate period of A and C and the number of subperiods is adjusted accordingly.

The iterative method (Algorithm 5) of [1] and its dual version are employed.

Reference:

[1] A. Varga. Periodic Lyapunov equations: some applications and new algorithms. Int. J. Control, vol, 67, pp, 69-87, 1997.

PeriodicSystems.prdplyap — Function

prdplyap(A, C) -> U

Compute the upper triangular factor U of the solution X = U'*U of the reverse time periodic discrete-time Lyapunov matrix equation

A'σXA + C'C = X

where σ is the forward shift operator σX(i) = X(i+1). The periodic matrix A must be stable, i.e., have all characteristic multipliers with moduli less than one.

The periodic matrices A and C must have the same type, the same dimensions and commensurate periods. The resulting upper triangular periodic matrix U has the period set to the least common commensurate period of A and C and the number of subperiods is adjusted accordingly.

Note: X is the observability Gramian of the periodic pair (A,C).

PeriodicSystems.pfdplyap — Function

pfdplyap(A, B) -> U

Compute the upper triangular factor U of the solution X = U*U' of the forward-time periodic discrete-time Lyapunov equation

AXA' + BB' = σX

where σ is the forward shift operator σX(i) = X(i+1). The periodic matrix A must be stable, i.e., have all characteristic multipliers with moduli less than one.

The periodic matrices A and B must have the same type, the same dimensions and commensurate periods. The resulting upper triangular periodic matrix U has the period set to the least common commensurate period of A and B and the number of subperiods is adjusted accordingly.

Note: X is the reachability Gramian of the periodic pair (A,B).