Periodic Riccati equation solvers

pcric Solution of periodic Riccati differential equations.

pgcric Computation of periodic generators for periodic Riccati differential equations.

prcric Solution of control-related reverse-time periodic Riccati differential equation.

pfcric Solution of filtering-related forward-time periodic Riccati differential equation.

prdric Solution of control-related reverse-time periodic Riccati difference equation.

pfdric Solution of filtering-related forward-time periodic Riccati difference equation.

PeriodicSystems.pcric — Function

pcric(A, R, Q; K = 10, adj = false, solver, reltol, abstol, fast) -> (X, EVALS)

Solve the periodic Riccati differential equation

.                                                 
X(t) = A(t)X(t) + X(t)A(t)' + Q(t) - X(t)R(t)X(t) ,  if adj = false,

or

 .                                                
-X(t) = A(t)'X(t) + X(t)A(t) + Q(t) - X(t)R(t)X(t) , if adj = true

and compute the stable closed-loop characteristic multipliers in EVALS (see pgcric for details).

The periodic matrices A, R and Q must have the same type, the same dimensions and commensurate periods, and additionally R and Q must be symmetric. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A, R and Q and the number of subperiods is adjusted accordingly.

If fast = true (default) the multiple-shooting method is used in conjunction with fast pencil reduction techniques, as proposed in [1], to determine the periodic solution in t = 0 and a multiple point generator of the appropriate periodic differential Riccati equation is determined (see [2] for details). If fast = false, the multiple-shooting method is used in conjunction with the periodic Schur decomposition to determine multiple point generators directly from the stable periodic invariant subspace of an appropriate symplectic transition matrix (see also [2] for more details).

The keyword argument K specifies the number of grid points to be used for the discretization of the continuous-time problem (default: K = 10). The multiple point generator is computed by solving the appropriate discrete-time periodic Lyapunov equation using the periodic Schur method of [2]. The resulting periodic time-series matrix is finally converted to the appropriate periodic representation.

The ODE solver to be employed to convert the continuous-time problem into a discrete-time problem can be specified using the keyword argument solver, together with the required relative accuracy reltol (default: reltol = 1.e-4) and absolute accuracy abstol (default: abstol = 1.e-7) (see tvstm). For large values of K, parallel computation of the matrices of the discrete-time problem can be alternatively performed by starting Julia with several execution threads. The number of execution threads is controlled either by using the -t/--threads command line argument or by using the JULIA_NUM_THREADS environment variable.

References

[1] A. Varga. On solving periodic differential matrix equations with applications to periodic system norms computation. Proc. IEEE CDC/ECC, Seville, 2005.

[2] A. Varga. On solving periodic Riccati equations. Numerical Linear Algebra with Applications, 15:809-835, 2008.

PeriodicSystems.pgcric — Function

pgcric(A, R, Q[, K = 1]; adj = false, solver, reltol, abstol, fast, PSD_SLICOT) -> (X, EVALS)

Compute periodic generators for the periodic Riccati differential equation in the filtering form

.                                                  
X(t) = A(t)X(t) + X(t)A(t)' + Q(t) - X(t)R(t)X(t), if adj = false,

or in the control form

 .                                              
-X(t) = A(t)'X(t) + X(t)A(t) + Q(t) - X(t)R(t)X(t) , if adj = true,

where A(t), R(t) and Q(t) are periodic matrices of commensurate periods, with A(t) square, R(t) symmetric and positive definite, and Q(t) symmetric and positive semidefinite. The resulting X is a collection of periodic generator matrices determined as a periodic time-series matrix with N components, where N = 1 if A(t), R(t) and Q(t) are constant matrices and N = K otherwise. EVALS contains the stable characteristic multipliers of the monodromy matrix of the corresponding Hamiltonian matrix (also called closed-loop characteristic multipliers). The period of X is set to the least common commensurate period of A(t), R(t) and Q(t) and the number of subperiods is adjusted accordingly. Any component matrix of X is a valid initial value to be used to generate the solution over a full period by integrating the appropriate differential equation. The multiple-shooting method proposed in [1] is employed, first, to convert the continuous-time periodic Lyapunov into a discrete-time periodic Lyapunov equation satisfied by the generator solution in K time grid points and then to compute the solution by solving an appropriate discrete-time periodic Lyapunov equation using the periodic Schur method of [2].

The ODE solver to be employed to convert the continuous-time problem into a discrete-time problem can be specified using the keyword argument solver, together with the required relative accuracy reltol (default: reltol = 1.e-4) and absolute accuracy abstol (default: abstol = 1.e-7) (see tvstm). For large values of K, parallel computation of the matrices of the discrete-time problem can be alternatively performed by starting Julia with several execution threads. The number of execution threads is controlled either by using the -t/--threads command line argument or by using the JULIA_NUM_THREADS environment variable.

References

[1] A. Varga. On solving periodic differential matrix equations with applications to periodic system norms computation. Proc. IEEE CDC/ECC, Seville, 2005.

[2] A. Varga. Periodic Lyapunov equations: some applications and new algorithms. Int. J. Control, vol, 67, pp, 69-87, 1997.

PeriodicSystems.prcric — Function

prcric(A, B, R, Q; K = 10, solver, reltol, abstol) -> (X, EVALS, F)

Compute the symmetric stabilizing solution X(t) of the periodic control related Riccati differential equation

 .                                                -1 
-X(t) = A(t)'X(t) + X(t)A(t) + Q(t) - X(t)B(t)R(t)  B(t)'X(t) ,

the periodic stabilizing state-feedback gain

           -1
F(t) = R(t)  B(t)'X(t)

and the corresponding stable characteristic multipliers EVALS of A(t)-B(t)F(t).

If fast = true (default) the multiple-shooting method is used in conjunction with fast pencil reduction techniques, while if fast = false, the multiple-shooting method is used in conjunction with the periodic Schur decomposition approach.

The periodic matrices A, B, R and Q must have the same type and commensurate periods, and additionally R must be symmetric positive definite and Q must be symmetric positive semidefinite. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A, B, R and Q and the number of subperiods is adjusted accordingly.

PeriodicSystems.pfcric — Function

pfcric(A, C, R, Q; K = 10, solver, reltol, abstol, fast) -> (X, EVALS, F)

Compute the symmetric stabilizing solution X(t) of the periodic filtering related Riccati differential equation

.                                                 -1 
X(t) = A(t)X(t) + X(t)A(t)' + Q(t) - X(t)C(t)'R(t)  C(t)X(t) ,

the periodic stabilizing Kalman gain

                    -1
F(t) = X(t)C(t)'R(t)

and the corresponding stable characteristic multipliers EVALS of A(t)-F(t)C(t).

If fast = true (default) the multiple-shooting method is used in conjunction with fast pencil reduction techniques, while if fast = false, the multiple-shooting method is used in conjunction with the periodic Schur decomposition approach.

The periodic matrices A, C, R and Q must have the same type and commensurate periods, and additionally R must be symmetric positive definite and Q must be symmetric positive semidefinite. The resulting symmetric periodic solution X has the period set to the least common commensurate period of A, C, R and Q and the number of subperiods is adjusted accordingly.

PeriodicSystems.prdric — Function

 prdric(A, B, R, Q[, S]; itmax = 0, nodeflate = false, fast, rtol) -> (X, EVALS, F)

Solve the periodic Riccati difference equation

  X(i) = Q(i) + A(i)'X(i+1)A(i) - (A(i)'X(i+1)B(i) + S(i))*
                                 -1
         (B(i)'X(i+1)B(i) + R(i))  (A(i)'X(i+1)B(i) + S(i))'

and compute the stabilizing periodic state feedback

                                  -1
  F(i) = -(B(i)'X(i+1)B(i) + R(i))  (B(i)'X(i+1)A(i) + S(i)')

and the corresponding stable closed-loop characteristic multipliers of A(i)-B(i)F(i) in EVALS.

The n×n and n×m periodic matrices A(i) and B(i) are contained in the PeriodicArray objects A and B, and must have the same sampling time. R(i), Q(i) and S(i) are m×m, n×n and n×m periodic matrices of same sampling times as A and B, and such that R(i) and Q(i) are symmetric. R(i), Q(i) and S(i) are contained in the PeriodicArray objects R, Q and S. R, Q and S can be alternatively provided as constant real matrices. The resulting symmetric periodic solution X and periodic state feedback gain F have the period set to the least common commensurate period of A, B, R and Q and the number of subperiods is adjusted accordingly.

If fast = true (default), the fast structure exploiting pencil reduction based method of [1] is used to determine a periodic generator in X(1), which allows to generate iteratively the solution over the whole period. If fast = false, the periodic Schur decomposition based approach of [1] is employed, applied to a symplectic pair of periodic matrices. If nodeflate = false (default), the underlying periodic pencil is preprocessed to eliminate (deflate) the infinite characteristic multipliers originating from the problem structure. If nodeflate = true, no preliminary deflation is performed.

An iterative refining of the accuracy of the computed solution can be performed by using itmax = k, with k > 0 (default: k = 0).

To detect singularities of involved matrices, the keyword parameter rtol = tol can be used to specify the lower bound for the 1-norm reciprocal condition number. The default value of tol is n*ϵ, where ϵ is the working machine epsilon.

References

[1] A. Varga. On solving periodic Riccati equations. Numerical Linear Algebra with Applications, 15:809-835, 2008.

 prdric(A, B, R, Q[, S]; itmax = 0, nodeflate = false, fast, rtol) -> (X, EVALS, F)

Solve the periodic Riccati difference equation

  X(i) = Q(i) + A(i)'X(i+1)A(i) - (A(i)'X(i+1)B(i) + S(i))*
                                 -1
         (B(i)'X(i+1)B(i) + R(i))  (A(i)'X(i+1)B(i) + S(i))'

and compute the stabilizing periodic state feedback

                                  -1
  F(i) = -(B(i)'X(i+1)B(i) + R(i))  (B(i)'X(i+1)A(i) + S(i)')

and the corresponding stable closed-loop core characteristic multipliers of A(i)-B(i)F(i) in EVALS.

The n(i+1)×n(i) and n(i+1)×m periodic matrices A(i) and B(i) are contained in the PeriodicMatrix objects A and B, and must have the same sampling time. R(i), Q(i) and S(i) are m×m, n(i)×n(i) and n(i)×m periodic matrices of same sampling times as A and B, and such that R(i) and Q(i) are symmetric. R(i), Q(i) and S(i) are contained in the PeriodicMatrix objects R, Q and S. R, Q and S can be alternatively provided as constant real matrices. The resulting n(i)×n(i) symmetric periodic solution X(i) and m×n(i) periodic state feedback gain F(i) have the period set to the least common commensurate period of A, B, R and Q and the number of subperiods is adjusted accordingly.

If fast = true (default), the fast structure exploiting pencil reduction based method of [1] is used to determine a periodic generator in X(j), which allows to generate iteratively the solution over the whole period. The value of j corresponds to the least dimension nc of n(i) (which is also the number of core characteristic multipliers). If fast = false, the periodic Schur decomposition based approach of [1] is employed, applied to a symplectic pair of periodic matrices. If nodeflate = false (default), the underlying periodic pencil is preprocessed to eliminate (deflate) the infinite characteristic multipliers originating from the problem structure. If nodeflate = true, no preliminary deflation is performed.

An iterative refining of the accuracy of the computed solution can be performed by using itmax = k, with k > 0 (default: k = 0).

To detect singularities of involved matrices, the keyword parameter rtol = tol can be used to specify the lower bound for the 1-norm reciprocal condition number. The default value of tol is n*ϵ, where ϵ is the working machine epsilon and n is the maximum of n(i).

References

[1] A. Varga. On solving periodic Riccati equations. Numerical Linear Algebra with Applications, 15:809-835, 2008.

PeriodicSystems.pfdric — Function

 pfdric(A, C, R, Q[, S]; itmax = 0, nodeflate = false, fast, rtol) -> (X, EVALS, F)

Solve the periodic Riccati difference equation

  X(i+1) = Q(i) + A(i)X(i)A(i)' - (A(i)X(i)C(i)' + S(i))*
                                 -1
           (C(i)X(i)C(i)' + R(i))  (A(i)X(i)C(i)' + S(i))'

and compute the stabilizing periodic Kalman gain

                                                       -1
  F(i) = -(C(i)X(i)A(i)' + S(i)')(C(i)X(i)C(i)' + R(i))

and the corresponding stable Kalman filter characteristic multipliers of A(i)-F(i)C(i) in EVALS.

The n×n and m×n periodic matrices A(i) and C(i) are contained in the PeriodicArray objects A and C, and must have the same sampling time. R(i), Q(i) and S(i) are m×m, n×n and m×n periodic matrices of same sampling times as A and C, and such that R(i) and Q(i) are symmetric. R, Q and S can be alternatively provided as constant real matrices. The resulting symmetric periodic solution X and Kalman filter gain F have the period set to the least common commensurate period of A, C, R and Q and the number of subperiods is adjusted accordingly.

The dual method of [1] is employed (see prdric for the description of keyword parameters).

References

[1] A. Varga. On solving periodic Riccati equations. Numerical Linear Algebra with Applications, 15:809-835, 2008.

 pfdric(A, C, R, Q[, S]; itmax = 0, nodeflate = false, fast, rtol) -> (X, EVALS, F)

Solve the periodic Riccati difference equation

  X(i+1) = Q(i) + A(i)X(i)A(i)' - (A(i)X(i)C(i)' + S(i))*
                                 -1
           (C(i)X(i)C(i)' + R(i))  (A(i)X(i)C(i)' + S(i))'

and compute the stabilizing periodic Kalman gain

                                                       -1
  F(i) = -(C(i)X(i)A(i)' + S(i)')(C(i)X(i)C(i)' + R(i))

and the corresponding stable Kalman filter core characteristic multipliers of A(i)-F(i)C(i) in EVALS.

The n(i+1)×n(i) and m×n(i) periodic matrices A(i) and C(i) are contained in the PeriodicMatrix objects A and C, and must have the same sampling time. R(i), Q(i) and S(i) are m×m, n(i)×n(i) and m×n(i) periodic matrices of same sampling times as A and C, and such that R(i) and Q(i) are symmetric. R, Q and S can be alternatively provided as constant real matrices. The resulting symmetric periodic solution X and Kalman filter gain F have the period set to the least common commensurate period of A, C, R and Q and the number of subperiods is adjusted accordingly.

The dual method of [1] is employed (see prdric for the description of keyword parameters).

References

[1] A. Varga. On solving periodic Riccati equations. Numerical Linear Algebra with Applications, 15:809-835, 2008.