Periodic system utilities

pseig Characteristic multipliers of a periodic matrix.

psceig Characteristic exponents of a periodic matrix.

psceighr Characteristic exponents of a periodic matrix in Harmonic Array representation.

psceigfr Characteristic exponents of a periodic matrix in Fourier Function Matrix representation.

monodromy Monodromy matrix of a linear periodic time-varying system of ODE.

tvstm State transition matrix of a linear time-varying system of ODE.

psreduc_reg Fast reduction of a lifted regular pencil corresponding to a product of matrices.

tvmeval Time response evaluation of a continuous-time periodic matrix.

hreval Evaluation of a harmonic array for a numerical or symbolic time value.

hrchop Removal of the negligible trailing terms of a harmonic representation.

hrtrunc Truncation of a harmonic representation.

pmaverage Evaluation of the time averaged matrix of a continuous-time periodic matrix.

pslyapd Solution of periodic discrete-time Lyapunov equations using periodic Schur decomposition.

pslyapd2 Solution of a pair periodic discrete-time Lyapunov equations using a single periodic Schur decomposition.

pslyapdkr Solution of periodic discrete-time Lyapunov equations using Kronecker product expansions.

psplyapd Solution of positve periodic discrete-time Lyapunov equations.

PeriodicSystems.pseig — Function

 pseig(A, K = 1; lifting = false, solver, reltol, abstol, dt) -> ev

Compute the characteristic multipliers of a continuous-time periodic matrix.

For the given square periodic matrix A(t) of period T, the characteristic multipliers ev are the eigenvalues of the monodromy matrix Ψ = Φ(T,0), where Φ(t,τ) is the state transition matrix satisfying the homogeneous linear ODE

dΦ(t,τ)/dt = A(t)Φ(t,τ),  Φ(τ,τ) = I.

If lifting = false, Ψ is computed as a product of K state transition matrices Ψ = Ψ_K*...*Ψ_1 (see monodromy with the associated keyword arguments). The eigenvalues are computed using the periodic Schur decomposition method of [1].

If lifting = true, Ψ is (implicitly) expressed as Ψ = inv(N)*M, where M-λN is a regular pencil with N invertible and the eigenvalues of M-λN are the same as those of the matrix product Ψ := Ψ_K*...*Ψ_1. An efficient version of the structure exploiting fast reduction method of [2] is employed, which embeds the determination of transition matrices into the reduction algorithm. This option may occasionally lead to inaccurate results for large values of K. A may be a PeriodicFunctionMatrix, or a PeriodicSymbolicMatrix, or a HarmonicArray or a PeriodicTimeSeriesMatrix.

References

[1] A. Bojanczyk, G. Golub, and P. Van Dooren, The periodic Schur decomposition. Algorithms and applications, Proc. SPIE 1996.

[2] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

 ev = pseig(A::PeriodicArray; rev = true, fast = false)

Compute the eigenvalues of a product of p square matrices A(p)...*A(2)*A(1), if rev = true (default) (also called characteristic multipliers) or of A(1)*A(2)...A(p) if rev = false, without evaluating the product. The matrices A(1), ..., A(p) are contained in the n×n×p array A such that the i-th matrix A(i) is contained in A[:,:,i]. If fast = false (default) then the eigenvalues are computed using an approach based on the periodic Schur decomposition [1], while if fast = true the structure exploiting reduction [2] of an appropriate lifted pencil is employed. This later option may occasionally lead to inaccurate results for large number of matrices.

References

[1] A. Bojanczyk, G. Golub, and P. Van Dooren, The periodic Schur decomposition. Algorithms and applications, Proc. SPIE 1996.

[2] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

 ev = pseig(A::PeriodicMatrix[, k = 1]; rev = true, fast = false)

Compute the eigenvalues of a square cyclic product of p matrices A(k-1)...*A(2)*A(1)*A(p)...*A(k), if rev = true (default) or A(k)*A(k+1)*...A(p)*A(1)...A(k-1) if rev = false, without evaluating the product. The argument k specifies the starting index (default: k = 1). The matrices A(1), ..., A(p) are contained in the p-vector of matrices A such that the i-th matrix A(i), of dimensions m(i)×n(i), is contained in A[i]. If fast = false (default) then the eigenvalues are computed using an approach based on the periodic Schur decomposition [1], while if fast = true the structure exploiting reduction [2] of an appropriate lifted pencil is employed. This later option may occasionally lead to inaccurate results for large number of matrices.

Note: The first nmin components of ev contains the core eigenvalues of the appropriate matrix product, where nmin is the minimum row dimensions of matrices A[i], for i = 1, ..., p, while the last ncur-nmin components of ev are zero, where ncur is the column dimension of A[k] if rev = true or the row dimension of A[k] if rev = false.

References

[1] A. Bojanczyk, G. Golub, and P. Van Dooren, The periodic Schur decomposition. Algorithms and applications, Proc. SPIE 1996.

[2] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

PeriodicSystems.psceig — Function

 psceig(A::PeriodicFunctionMatrix[, K = 1]; lifting = false, solver, reltol, abstol, dt) -> ce

Compute the characteristic exponents of a periodic matrix.

For a given square continuous-time periodic function matrix A(t) of period T, the characteristic exponents ce are computed as log.(ev)/T, where ev are the characteristic multipliers (i.e., the eigenvalues of the monodromy matrix of A(t)). For available options see pseig(::PeriodicFunctionMatrix). For a given square discrete-time periodic matrix A(t) of discrete period N, the characteristic exponents ce are computed as ev.^-N.

psceig(A::AbstractPeriodicArray[, k]; kwargs...) -> ce

Compute the characteristic exponents of a cyclic matrix product of p matrices.

The characteristic exponents of a product of p matrices are computed as the pth roots of the characteristic multipliers. These are computed as the eigenvalues of the square cyclic product of p matrices A(k-1)...*A(2)*A(1)*A(p)...*A(k), if rev = true (default) or A(k)*A(k+1)*...A(p)*A(1)...A(k-1) if rev = false, without evaluating the product. The argument k specifies the starting index (default: k = 1). The matrices A(1), ..., A(p) are contained in the p-vector of matrices A such that the i-th matrix A(i), of dimensions m(i)×n(i), is contained in A[i]. The keyword arguments kwargs are those of pseig(::PeriodicMatrix).

Note: The first nmin components of ce contains the core characteristic exponents of the appropriate matrix product, where nmin is the minimum row dimensions of matrices A[i], for i = 1, ..., p, while the last components of ce are zero.

PeriodicSystems.psceighr — Function

psceighr(Ahr::HarmonicArray[, N]; P, nperiod, shift, atol) -> ce

Compute the characteristic exponents of a continuous-time periodic matrix in Harmonic Array representation.

For a given square continuous-time periodic function matrix Ahr(t) of period T in a Harmonic Array representation, the characteristic exponents ce are computed as the eigenvalues of a truncated harmonic state operator A(N)-E(N) lying in the fundamental strip -ω/2 < Im(λ) ≤ ω/2, where ω = 2π/T. If Ahr(t) has the harmonic components A_0, A_1, ..., A_p, then for N ≥ p, P = 1 and nperiod = 1, the matrices A(N) and E(N) are built as

       ( A_0  A_{-1} …  A_{-p}        0    )           ( -im*ϕ_{-N}I                                 0        )
       ( A_1   A_0             ⋱           )           (     ⋮       ⋱                                        )
       (  ⋮         ⋱            ⋱         )           (               -im*ϕ_{-1}*I                           )
A(N) = ( A_p             ⋱          A_{-p} ) , E(N) =  (                           -im*ϕ_{0}*I                )
       (        ⋱           ⋱         ⋮    )           (     ⋮                                  ⋱              )
       (  0        A_p      …         A_0  )           (     0                                   -im*ϕ_{N}I   )

with ϕ_{i} := shift+i*ω. If N < p, then a truncated full block Toeplitz matrix A(N) is built using the first N harmonic components. The default value used for N is N = max(10,p-1).

Generally, for given P ≥ 1 and nperiod ≥ 1, the block Toeplitz matrix A(N) (and also E(N)) is constructed with (2N*np+1)×(2N*np+1) blocks, with np = P*nperiod, such that each A_i is preceeded in its column by np-1 zero blocks, each A_{-i} is preceeded in its row by np-1 zero blocks and all diagonal blocks are equal toA_0.

The keyword argument atol (default: atol = 1.e-10) is a tolerance on the magnitude of the trailing components of the associated eigenvectors used to validate their asymptotic (exponential) decay. Only eigenvalues satisfying this check are returned in ce.

References

[1] J. Zhou, T. Hagiwara, and M. Araki. Spectral characteristics and eigenvalues computation of the harmonic state operators in continuous-time periodic systems. Systems and Control Letters, 53:141–155, 2004.

PeriodicSystems.psceigfr — Function

psceigfr(A::FourierFunctionMatrix[, N]; P, atol) -> ce

Compute the characteristic exponents of a continuous-time periodic matrix in Fourier Function Matrix representation.

For a given square continuous-time periodic function matrix A(t) of period T in a Fourier Function Matrix representation, the characteristic exponents ce are computed as the eigenvalues of the state operator A(t)-D*I lying in the fundamental strip -ω/2 < Im(λ) ≤ ω/2, where ω = 2π/T. A finite dimensional truncated matrix of order n*(2*N*P+1) is built to approximate A(t)-D*I, where n is the order of A(t), N is the number of selected harmonic components in the Fourier representation and P is the period multiplication number (default: P = 1). The default value used for N is N = max(10,p-1), where p the number of harmonics terms of A(t) (see FourierFunctionMatrix).

The keyword argument atol (default: atol = 1.e-10) is a tolerance on the magnitude of the trailing components of the associated eigenvectors used to validate their asymptotic (exponential) decay. Only eigenvalues satisfying this check are returned in ce.

PeriodicSystems.monodromy — Function

 monodromy(A[, K = 1]; solver, reltol, abstol, dt) -> Ψ::PeriodicArray

Compute the monodromy matrix for a linear ODE with periodic time-varying coefficients.

For the given square periodic matrix A(t) of period T and subperiod T′ = T/k, where k is the number of subperiods, the monodromy matrix Ψ = Φ(T′,0) is computed, where Φ(t,τ) is the state transition matrix satisfying the homogeneous linear ODE

dΦ(t,τ)/dt = A(t)Φ(t,τ),  Φ(τ,τ) = I.

A(t) can be specified as a PeriodicFunctionMatrix, HarmonicArray or FourierFunctionMatrix.

If K > 1, then Ψ = Φ(T′,0) is determined as a product of K matrices Ψ = Ψ_K*...*Ψ_1, where for Δ := T′/K, Ψ_i = Φ(iΔ,(i-1)Δ) is the state transition matrix on the time interval [(i-1)Δ,iΔ].

The state transition matrices Φ(iΔ,(i-1)Δ) are computed by integrating numerically the above homogeneous linear ODE. The ODE solver to be employed can be specified using the keyword argument solver, together with the required relative accuracy reltol (default: reltol = 1.e-3), absolute accuracy abstol (default: abstol = 1.e-7) and/or the fixed step length dt (default: dt = min(Δ, Δ*K′/100)) (see tvstm). For large values of K, parallel computation of factors can be alternatively performed by starting Julia with several execution threads. The number of execution threads is controlled either by using the -t/--threads command line argument or by using the JULIA_NUM_THREADS environment variable.

PeriodicSystems.tvstm — Function

 tvstm(A, tf, t0; solver, reltol, abstol, dt) -> Φ

Compute the state transition matrix for a linear ODE with periodic time-varying coefficients. For the given periodic square matrix A(t), initial time t0 and final time tf, the state transition matrix Φ(tf,t0) is computed by integrating numerically the homogeneous linear ODE

  dΦ(t,t0)/dt = A(t)Φ(t,t0),  Φ(t0,t0) = I

on the time interval [t0,tf]. A(t) can be specified as a PeriodicFunctionMatrix, HarmonicArray or FourierFunctionMatrix.

The ODE solver to be employed can be specified using the keyword argument solver (see below), together with the required relative accuracy reltol (default: reltol = 1.e-3), absolute accuracy abstol (default: abstol = 1.e-7) and/or the fixed step length dt (default: dt = tf-t0). Depending on the desired relative accuracy reltol, lower order solvers are employed for reltol >= 1.e-4, which are generally very efficient, but less accurate. If reltol < 1.e-4, higher order solvers are employed able to cope with high accuracy demands.

The following solvers from the OrdinaryDiffEq.jl package can be selected:

solver = "non-stiff" - use a solver for non-stiff problems (Tsit5() or Vern9());

solver = "stiff" - use a solver for stiff problems (Rodas4() or KenCarp58());

solver = "linear" - use a special solver for linear ODEs (MagnusGL6()) with fixed time step dt;

solver = "symplectic" - use a symplectic Hamiltonian structure preserving solver (IRKGL16());

solver = "" - use the default solver, which automatically detects stiff problems (AutoTsit5(Rosenbrock23()) or AutoVern9(Rodas5())).

PeriodicSystems.psreduc_reg — Function

psreduc_reg(A) -> (M, N)

Determine for a n×n×p array A, the matrix pair (M, N) with N invertible and M-λN regular, such that the eigenvalues of M-λN are the same as those of the matrix product A(p)*A(p-1)*...*A(1), where A(i) is contained in A[:,:,i]. The structure exploiting fast reduction method of [1] is employed to determine M and N.

[1] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

psreduc_reg(A,E) -> (M, N)

Determine for a pair of n×n×p arrays (A,E), the matrix pair (M, N) with M-λN regular, such that the eigenvalues of M-λN are the same as those of the quotient matrix product inv(E(p))*(A(p)*inv(E(p-1))*A(p-1)*...*inv(E(1))*A(1), where A(i) is contained in A[:,:,i] and E(i) is contained in E[:,:,i]. The structure exploiting fast reduction method of [1] is employed to determine M and N.

[1] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

psreduc_reg(A) -> (M, N)

Determine for a p-dimensional vector of rectangular matrices A, the matrix pair (M, N) with N invertible and M-λN regular, such that the eigenvalues of M-λN are the same as those of the square matrix product A(p)*A(p-1)*...*A(1), where A(i) is contained in A[i]. The structure exploiting fast reduction method of [1] is employed to determine M and N.

[1] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

PeriodicSystems.tvmeval — Function

 tvmeval(At::PeriodicTimeSeriesMatrix, t; method = "linear") -> A::Vector{Matrix}

Evaluate the time response of a periodic time series matrix.

For the periodic time series matrix At and the vector of time values t, an interpolation/extrapolation based approximation A[i] is evaluated for each time value t[i]. The keyword parameter method specifies the interpolation/extrapolation method to be used for periodic data. The following interpolation methods from the Interpolations.jl package can be selected:

method = "constant" - use periodic B-splines of degree 0;

method = "linear" - use periodic B-splines of degree 1 (periodic linear interpolation) (default);

method = "quadratic" - use periodic B-splines of degree 2 (periodic quadratic interpolation);

method = "cubic" - use periodic B-splines of degree 3 (periodic cubic interpolation).

 tvmeval(Ahr::HarmonicArray, t; ntrunc, exact = true) -> A::Vector{Matrix}

Evaluate the time response of a harmonic array.

For the harmonic array Ahr representing representing a continuous-time time periodic matrix A(t) and the vector of time values t, A[i] = A(t[i]) is computed for each time value t[i]. If exact = true (default) an exact evaluation is computed, while for exact = false, a linear interpolation based approximation is computed (potentially more accurate in intersample points). The keyword argument ntrunc specifies the number of harmonics to be used for evaluation (default: maximum possible number of harmonics).

 tvmeval(Asym::PeriodicSymbolicMatrix, t) -> A::Vector{Matrix}

Evaluate the time response of a periodic symbolic matrix.

For the periodic symbolic matrix Asym representing a time periodic matrix A(t) and the vector of time values t, A[i] = A(t[i]) is evaluated for each time value t[i].

 tvmeval(Af::PeriodicFunctionMatrix, t) -> A::Vector{Matrix}

Evaluate the time response of a periodic function matrix.

For the periodic function matrix Af representing a time periodic matrix A(t) and the vector of time values t, A[i] = A(t[i]) is evaluated for each time value t[i].

 tvmeval(A::FourierFunctionMatrix, t) -> Aval::Vector{Matrix}

Evaluate the time response of a periodic function matrix.

For the periodic matrix A(t), in a Fourier Function Matrix representation, and the vector of time values t, Aval[i] = A(t[i]) is evaluated for each time value t[i].

PeriodicSystems.hreval — Function

 hreval(Ahr::HarmonicArray, t; ntrunc, exact = true) -> A::Matrix

Evaluate the harmonic array Ahr representing a continuous-time time periodic matrix A(t) for a numerical or symbolic time value t. For a real value t, if exact = true (default) an exact evaluation is computed, while for exact = false, a linear interpolation based approximation is computed (potentially more accurate in intersample points). The keyword argument ntrunc specifies the number of harmonics to be used for the evaluation (default: maximum possible number). If t is a symbolic variable, a symbolic evaluation of A(t) is performed (see also hr2psm)

PeriodicSystems.hrchop — Function

 hrchop(Ahr::HarmonicArray; tol) -> Ahrtrunc::HarmonicArray

Remove the trailing terms of a harmonic representation by deleting those whose norms are below a certain tolerance.

PeriodicSystems.hrtrunc — Function

 hrtrunc(Ahr::HarmonicArray, n) -> Ahrtrunc::HarmonicArray

Truncate a harmonic representation by deleting the trailing terms whose indices exceed certain number n of harmonics.

PeriodicSystems.pmaverage — Function

pmaverage(A) -> Am

Compute for the continuous-time periodic matrix A(t) the corresponding time averaged matrix Am over one period.

PeriodicSystems.pslyapd — Function

pslyapd(A, C; adj = true, stability_check = false) -> X

Solve the periodic discrete-time Lyapunov equation.

For the square n-th order periodic matrices A(i), i = 1, ..., pa and C(i), i = 1, ..., pc of periods pa and pc, respectively, the periodic solution X(i), i = 1, ..., p of period p = lcm(pa,pc) of the periodic Lyapunov equation is computed:

A(i)'*X(i+1)*A(i) + C(i) = X(i), i = 1, ..., p     for `adj = true`; 

A(i)*X(i)*A(i)' + C(i) = X(i+1), i = 1, ..., p     for `adj = false`.

The periodic matrices A and C are stored in the n×n×pa and n×n×pc 3-dimensional arrays A and C, respectively, and X results as a n×n×p 3-dimensional array.

Alternatively, the periodic matrices A and C can be stored in the pa- and pc-dimensional vectors of matrices A and C, respectively, and X results as a p-dimensional vector of matrices.

If stability_check = true, the stability of characteristic multipliers of A is checked and an error is issued if any characteristic multiplier has modulus equal to or larger than one.

The periodic discrete analog of the Bartels-Stewart method based on the periodic Schur form of the periodic matrix A is employed [1].

Reference:

[1] A. Varga. Periodic Lyapunov equations: some applications and new algorithms. Int. J. Control, vol, 67, pp, 69-87, 1997.

PeriodicSystems.pslyapd2 — Function

pslyapd2(A, C, E; stability_check = false) -> (X, Y)

Solve a pair of periodic discrete-time Lyapunov equations.

For the square n-th order periodic matrices A(i), i = 1, ..., pa, C(i), i = 1, ..., pc, and E(i), i = 1, ..., pe of periods pa, pc and pe, respectively, the periodic solutions X(i), i = 1, ..., p and Y(i), i = 1, ..., p of period p = lcm(pa,pc,pe) of the periodic Lyapunov equations are computed:

A(i)*X(i)*A(i)' + C(i) = X(i+1), i = 1, ..., p ,  
 
A(i)'*Y(i+1)*A(i) + E(i) = Y(i), i = 1, ..., p .

The periodic matrices A, C and E are stored in the n×n×pa, n×n×pc and n×n×pe 3-dimensional arrays A, C and E, respectively, and X and Y result as n×n×p 3-dimensional arrays.

Alternatively, the periodic matrices A, C and E can be stored in the pa-, pc- and pe-dimensional vectors of matrices A, C and E, respectively, and X and Y result as p-dimensional vectors of matrices.

If stability_check = true, the stability of characteristic multipliers of A is checked and an error is issued if any characteristic multiplier has modulus equal to or larger than one.

The periodic discrete analog of the Bartels-Stewart method based on the periodic Schur form of the periodic matrix A is employed [1].

Reference:

[1] A. Varga. Periodic Lyapunov equations: some applications and new algorithms. Int. J. Control, vol, 67, pp, 69-87, 1997.

PeriodicSystems.pslyapdkr — Function

 pslyapdkr(A, C; adj = true) -> X

Solve the periodic discrete-time Lyapunov matrix equation

  A'σXA + C = X, if adj = true,

or

  A*X*A' + C =  σX, if adj = false,

where σ is the forward shift operator σX(i) = X(i+1). The periodic matrix A must not have characteristic multipliers on the unit circle. The periodic matrices A and C are either stored as 3-dimensional arrays or as as vectors of matrices.

The Kronecker product expansion of equations is employed and therefore this function is not recommended for large order matrices or large periods.

PeriodicSystems.psplyapd — Function

 psplyapd(A, C; adj = true, rtol = ϵ^(3/4)) -> U

Compute the upper triangular factor U of the solution X = U'U of the periodic discrete-time Lyapunov matrix equation

  A'σXA + C'C = X, if adj = true,

or of the solution X = UU' of the periodic discrete-time Lyapunov matrix equation

  AXA' + CC' =  σX, if adj = false,

where σ is the forward shift operator σX(i) = X(i+1). The periodic matrix A must be stable, i.e., have all characteristic multipliers with moduli less than one.

The periodic matrices A and C are either stored as 3-dimensional arrays or as as vectors of matrices.

The iterative method (Algorithm 5) of [1] and its dual version are employed.

Reference:

[1] A. Varga, "Periodic Lyapunov equations: some applications and new algorithms", Int. J. Control, vol. 67, pp. 69-87, 1997.