Periodic system utilities

pseig Characteristic multipliers of a periodic matrix.

psceig Characteristic exponents of a periodic matrix.

pschur Periodic Schur decomposition of a product of matrices.

phess Periodic Hessenberg decomposition of a product of matrices.

monodromy Monodromy matrix of a linear periodic time-varying system of ODE.

tvstm State transition matrix of a linear time-varying system of ODE.

psreduc_reg Fast reduction of a lifted regular pencil corresponding to a product of matrices.

tvmeval Time response evaluation of a continuous-time periodic matrix.

hreval Evaluation of a harmonic array for a numerical or symbolic time value.

pmaverage Evaluation of the time averaged matrix of a continuous-time periodic matrix.

PeriodicSystems.pseig — Function

 pseig(A, K = 1; lifting = false, solver, reltol, abstol, dt) -> ev

Compute the characteristic multipliers of a continuous-time periodic matrix.

For the given square periodic matrix A(t) of period T, the characteristic multipliers ev are the eigenvalues of the monodromy matrix Ψ = Φ(T,0), where Φ(t,τ) is the state transition matrix satisfying the homogeneous linear ODE

dΦ(t,τ)/dt = A(t)Φ(t,τ),  Φ(τ,τ) = I.

If lifting = false, Ψ is computed as a product of K state transition matrices Ψ = Ψ_K*...*Ψ_1 (see monodromy with the associated keyword arguments). The eigenvalues are computed using the periodic Schur decomposition method of [1].

If lifting = true, Ψ is (implicitly) expressed as Ψ = inv(N)*M, where M-λN is a regular pencil with N invertible and the eigenvalues of M-λN are the same as those of the matrix product Ψ := Ψ_K*...*Ψ_1. An efficient version of the structure exploiting fast reduction method of [2] is employed, which embeds the determination of transition matrices into the reduction algorithm. This option may occasionally lead to inaccurate results for large values of K. A may be a PeriodicFunctionMatrix, or a PeriodicSymbolicMatrix, or a HarmonicArray or a PeriodicTimeSeriesMatrix.

References

[1] A. Bojanczyk, G. Golub, and P. Van Dooren, The periodic Schur decomposition. Algorithms and applications, Proc. SPIE 1996.

[2] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

 ev = pseig(A::PeriodicMatrix; rev = true, fast = false)

Compute the eigenvalues of a product of p square matrices A(p)...*A(2)*A(1), if rev = true (default) (also called characteristic multipliers) or of A(1)*A(2)...A(p) if rev = false, without evaluating the product. The matrices A(1), ..., A(p) are contained in the n×n×p array A such that the i-th matrix A(i) is contained in A[:,:,i]. If fast = false (default) then the eigenvalues are computed using an approach based on the periodic Schur decomposition [1], while if fast = true the structure exploiting reduction [2] of an appropriate lifted pencil is employed. This later option may occasionally lead to inaccurate results for large number of matrices.

References

[1] A. Bojanczyk, G. Golub, and P. Van Dooren, The periodic Schur decomposition. Algorithms and applications, Proc. SPIE 1996.

[2] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

 ev = pseig(A::PeriodicArray; rev = true, fast = false, istart = 1)

Compute the eigenvalues of a square cyclic product of p matrices A(k-1)...*A(2)*A(1)*A(p)...*A(k), if rev = true (default) or A(k)*A(k+1)*...A(p)*A(1)...A(k-1) if rev = false, without evaluating the product. The keyword argument istart = k specifies the starting index (default: k = 1). The matrices A(1), ..., A(p) are contained in the p-vector of matrices A such that the i-th matrix A(i), of dimensions m(i)×n(i), is contained in A[i]. If fast = false (default) then the eigenvalues are computed using an approach based on the periodic Schur decomposition [1], while if fast = true the structure exploiting reduction [2] of an appropriate lifted pencil is employed. This later option may occasionally lead to inaccurate results for large number of matrices.

Note: The first nmin components of ev contains the core eigenvalues of the appropriate matrix product, where nmin is the minimum row dimensions of matrices A[i], for i = 1, ..., p, while the last ncur-nmin components of ev are zero, where ncur is the column dimension of A[k] if rev = true or the row dimension of A[k] if rev = false.

References

[1] A. Bojanczyk, G. Golub, and P. Van Dooren, The periodic Schur decomposition. Algorithms and applications, Proc. SPIE 1996.

[2] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

PeriodicSystems.psceig — Function

 pcseig(A, K = 1; lifting = false, solver, reltol, abstol, dt) -> ce

Compute the characteristic exponents of a periodic matrix. For a given square continuous-time periodic function matrix A(t) of period T, the characteristic exponents ce are computed as log.(ev)/T, where ev are the characteristic multipliers (i.e., the eigenvalues of the monodromy matrix of A(t)). For available options see pseig(::PeriodicFunctionMatrix). For a given square discrete-time periodic matrix A(t) of discrete period N, the characteristic exponents ce are computed as ev.^-N.

pcseig(Ahr::HarmonicArray[, N]; P, nperiod, shift, atol) -> ce

Compute the characteristic exponents of a continuous-time periodic matrix in Harmonic Array representation.

For a given square continuous-time periodic function matrix Ahr(t) of period T in a Harmonic Array representation, the characteristic exponents ce are computed as the eigenvalues of a truncated harmonic state operator A(N)-E(N) lying in the fundamental strip -ω/2 < Im(λ) ≤ ω/2, where ω = 2π/T. If Ahr(t) has the harmonic components A_0, A_1, ..., A_p, then for N ≥ p, P = 1 and nperiod = 1, the matrices A(N) and E(N) are built as

       ( A_0  A_{-1} …  A_{-p}        0    )           ( -im*ϕ_{-N}I                                 0        )
       ( A_1   A_0             ⋱           )           (     ⋮       ⋱                                        )
       (  ⋮         ⋱            ⋱         )           (               -im*ϕ_{-1}*I                           )
A(N) = ( A_p             ⋱          A_{-p} ) , E(N) =  (                           -im*ϕ_{0}*I                )
       (        ⋱           ⋱         ⋮    )           (     ⋮                                  ⋱              )
       (  0        A_p      …         A_0  )           (     0                                   -im*ϕ_{N}I   )

with ϕ{i} := shift+i*ω. If N < p, then a truncated _full block Toeplitz matrix A(N) is built using the first N harmonic components.

Generally, for given P ≥ 1 and nperiod ≥ 1, the block Toeplitz matrix A(N) (and also E(N)) is constructed with (2N*np+1)×(2N*np+1) blocks, with np = P*nperiod, such that each A_i is preceeded in its column by np-1 zero blocks, each A_{-i} is preceeded in its row by np-1 zero blocks and all diagonal blocks are equal toA_0.

The keyword argument atol (default: atol = 1.e-10) is a tolerance on the magnitude of the trailing components of the associated eigenvectors used to validate their asymptotic (exponential) decay. Only eigenvalues satisfying this check are returned in ce.

References

[1] J. Zhou, T. Hagiwara, and M. Araki. Spectral characteristics and eigenvalues computation of the harmonic state operators in continuous-time periodic systems. Systems and Control Letters, 53:141–155, 2004.

pcseig(A::FourierFunctionMatrix[, N]; P, atol) -> ce

Compute the characteristic exponents of a continuous-time periodic matrix in Fourier Function Matrix representation.

For a given square continuous-time periodic function matrix A(t) of period T in a Fourier Function Matrix representation, the characteristic exponents ce are computed as the eigenvalues of the state operator A(t)-D*I lying in the fundamental strip -ω/2 < Im(λ) ≤ ω/2, where ω = 2π/T. A finite dimensional truncated matrix of order n*(2*N*P+1) is built to approximate A(t)-D*I, where n is the order of A(t), N is the number of selected harmonic components in the Fourier representation (see FourierFunctionMatrix) and P is the period multiplication number (default: P = 1).

The keyword argument atol (default: atol = 1.e-10) is a tolerance on the magnitude of the trailing components of the associated eigenvectors used to validate their asymptotic (exponential) decay. Only eigenvalues satisfying this check are returned in ce.

PeriodicSystems.pschur — Function

 pschur(A; rev = true, withZ = true) -> (S, Z, ev, ischur, α, γ)

Compute the extended periodic Schur decomposition of a square product of matrices A(p)*...*A(2)*A(1), if rev = true (default) or A(1)*A(2)*...*A(p) if rev = false, without evaluating the product. The matrices A(1), ..., A(p) are contained in the p-vector of matrices A such that the i-th matrix A(i), of dimensions m(i)×n(i), is contained in A[i]. The resulting p-vectors S and Z contain the matrices S(1), ..., S(p) and the orthogonal matrices Z(1), ..., Z(p), respectively, such that for rev = true

       Z(2)' * A(1) * Z(1) = S(1),
       Z(3)' * A(2) * Z(2) = S(2),
              ...
       Z(1)' * A(p) * Z(p) = S(p),

and for rev = false

       Z(1)' * A(1) * Z(2) = S(1),
       Z(2)' * A(2) * Z(3) = S(2),
              ...
       Z(p)' * A(p) * Z(1) = S(p).

The resulting index ischur is determined such that m(ischur) ≤ m(i), ∀i. The resulting S(i), for i = 1, ..., p are in an extended periodic Schur form, with S(ischur) in a quasi-upper trapezoidal form and S(i) upper trapezoidal for i $\neq$ ischur. S(i) and Z(i) are contained in S[i] and Z[i], respectively. The first nmin components of ev := α .* γ contain the core eigenvalues of the appropriate matrix product, where nmin = m(ischur), while the last nmax-nmin components of ev are zero, where nmax is the largest row or column dimension of A(i), for i = 1, ..., p. The eigenvalues can be alternatively expressed as α .* γ, where γ contains suitable scaling parameters to avoid overflows or underflows in the expressions of the eigenvalues. The performed orthogonal transformations are not accumulated if withZ = false, in which case Z = nothing.

 pschur(A; sind = 1, rev = true, withZ = true) -> (S, Z, ev, ischur, α, γ)

Compute the Schur decomposition of a product of square matrices A(p)*...*A(2)*A(1), if rev = true (default) or A(1)*A(2)*...*A(p) if rev = false, without evaluating the product. The matrices A(1), ..., A(p) are contained in the n×n×p array A such that the i-th matrix A(i) is contained in A[:,:,i]. The resulting n×n×p arrays S and Z contain the matrices S(1), ..., S(p) and the orthogonal matrices Z(1), ..., Z(p), respectively, such that for rev = true

       Z(2)' * A(1) * Z(1) = S(1),
       Z(3)' * A(2) * Z(2) = S(2),
              ...
       Z(1)' * A(p) * Z(p) = S(p),

and for rev = false

       Z(1)' * A(1) * Z(2) = S(1),
       Z(2)' * A(2) * Z(3) = S(2),
              ...
       Z(p)' * A(p) * Z(1) = S(p).

If sind = ischur, with 1 ≤ ischur ≤ p (default ischur = 1), then S(i), for i = 1, ..., p are in a periodic Schur form, with S(ischur) in quasi-upper triangular (or Schur) form and S(i) upper triangular for i $\neq$ ischur. S(i) and Z(i) are contained in S[:,:,i] and Z[:,:,i], respectively. The vector ev contains the eigenvalues of the appropriate matrix product. The eigenvalues can be alternatively expressed as α .* γ, where γ contains suitable scaling parameters to avoid overflows or underflows in the expressions of the eigenvalues. The performed orthogonal transformations are not accumulated if withZ = false, in which case Z = nothing.

PeriodicSystems.phess — Function

 phess(A; hind = 1, rev = true, withZ = true) -> (H, Z, ihess)

Compute the Hessenberg decomposition of a product of square matrices A(p)*...*A(2)*A(1), if rev = true (default) or A(1)*A(2)*...*A(p) if rev = false, without evaluating the product. The matrices A(1), ..., A(p) are contained in the n×n×p array A such that the i-th matrix A(i) is contained in A[:,:,i]. The resulting n×n×p arrays H and Z contain the matrices H(1), ..., H(p) and the orthogonal matrices Z(1), ..., Z(p), respectively, such that for rev = true

       Z(2)' * A(1) * Z(1) = H(1),
       Z(3)' * A(2) * Z(2) = H(2),
              ...
       Z(1)' * A(p) * Z(p) = H(p),

and for rev = false

       Z(1)' * A(1) * Z(2) = H(1),
       Z(2)' * A(2) * Z(3) = H(2),
              ...
       Z(p)' * A(p) * Z(1) = H(p).

If hind = ihess, with 1 ≤ ihess ≤ p (default ihess = 1), then H(i), i = 1, ..., p are in a periodic Hessenberg form, with H(ihess) in upper Hessenberg form and H(i) upper triangular for i $\neq$ ihess. H(i) and Z(i) are contained in H[:,:,i] and Z[:,:,i], respectively. The performed orthogonal transformations are not accumulated if withZ = false, in which case Z = nothing.

PeriodicSystems.monodromy — Function

 monodromy(A::PeriodicFunctionMatrix, K = 1; solver, reltol, abstol, dt) -> Ψ::PeriodicArray

Compute the monodromy matrix for a linear ODE with periodic time-varying coefficients.

For the given square periodic function matrix A(t) of period T and subperiod T′ = T/k, where k is the number of subperiods, the monodromy matrix Ψ = Φ(T′,0) is computed, where Φ(t,τ) is the state transition matrix satisfying the homogeneous linear ODE

dΦ(t,τ)/dt = A(t)Φ(t,τ),  Φ(τ,τ) = I.

If K > 1, then Ψ = Φ(T′,0) is determined as a product of K matrices Ψ = Ψ_K*...*Ψ_1, where for Δ := T′/K, Ψ_i = Φ(iΔ,(i-1)Δ) is the state transition matrix on the time interval [(i-1)Δ,iΔ].

The state transition matrices Φ(iΔ,(i-1)Δ) are computed by integrating numerically the above homogeneous linear ODE. The ODE solver to be employed can be specified using the keyword argument solver, together with the required relative accuracy reltol (default: reltol = 1.e-3), absolute accuracy abstol (default: abstol = 1.e-7) and/or the fixed step length dt (default: dt = min(Δ, Δ*K′/100)) (see tvstm). For large values of K, parallel computation of factors can be alternatively performed by starting Julia with several execution threads. The number of execution threads is controlled either by using the -t/--threads command line argument or by using the JULIA_NUM_THREADS environment variable.

PeriodicSystems.tvstm — Function

 tvstm(A::PeriodicFunctionMatrix, tf, t0; solver, reltol, abstol, dt) -> Φ

Compute the state transition matrix for a linear ODE with time-varying coefficients. For the given time-dependent square matrix function A(t), initial time t0 and final time tf, the state transition matrix Φ(tf,t0) is computed by integrating numerically the homogeneous linear ODE

  dΦ(t,t0)/dt = A(t)Φ(t,t0),  Φ(t0,t0) = I

on the time interval [t0,tf]. The ODE solver to be employed can be specified using the keyword argument solver (see below), together with the required relative accuracy reltol (default: reltol = 1.e-3), absolute accuracy abstol (default: abstol = 1.e-7) and/or the fixed step length dt (default: dt = tf-t0). Depending on the desired relative accuracy reltol, lower order solvers are employed for reltol >= 1.e-4, which are generally very efficient, but less accurate. If reltol < 1.e-4, higher order solvers are employed able to cope with high accuracy demands.

The following solvers from the OrdinaryDiffEq.jl package can be selected:

solver = "non-stiff" - use a solver for non-stiff problems (Tsit5() or Vern9());

solver = "stiff" - use a solver for stiff problems (Rodas4() or KenCarp58());

solver = "linear" - use a special solver for linear ODEs (MagnusGL6()) with fixed time step dt;

solver = "symplectic" - use a symplectic Hamiltonian structure preserving solver (IRKGL16());

solver = "" - use the default solver, which automatically detects stiff problems (AutoTsit5(Rosenbrock23()) or AutoVern9(Rodas5())).

PeriodicSystems.psreduc_reg — Function

psreduc_reg(A) -> (M, N)

Determine for a n×n×p array A, the matrix pair (M, N) with N invertible and M-λN regular, such that the eigenvalues of M-λN are the same as those of the matrix product A(p)*A(p-1)*...*A(1), where A(i) is contained in A[:,:,i]. The structure exploiting fast reduction method of [1] is employed to determine M and N.

[1] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

psreduc_reg(A) -> (M, N)

Determine for a p-dimensional vector of rectangular matrices A, the matrix pair (M, N) with N invertible and M-λN regular, such that the eigenvalues of M-λN are the same as those of the square matrix product A(p)*A(p-1)*...*A(1), where A(i) is contained in A[i]. The structure exploiting fast reduction method of [1] is employed to determine M and N.

[1] A. Varga & P. Van Dooren. Computing the zeros of periodic descriptor systems. Systems and Control Letters, 50:371-381, 2003.

PeriodicSystems.tvmeval — Function

 tvmeval(At::PeriodicTimeSeriesMatrix, t; method = "linear") -> A::Vector{Matrix}

Evaluate the time response of a periodic time series matrix.

For the periodic time series matrix At and the vector of time values t, an interpolation/extrapolation based approximation A[i] is evaluated for each time value t[i]. The keyword parameter method specifies the interpolation/extrapolation method to be used for periodic data. The following interpolation methods from the Interpolations.jl package can be selected:

method = "constant" - use periodic B-splines of degree 0;

method = "linear" - use periodic B-splines of degree 1 (periodic linear interpolation) (default);

method = "quadratic" - use periodic B-splines of degree 2 (periodic quadratic interpolation);

method = "cubic" - use periodic B-splines of degree 3 (periodic cubic interpolation).

 tvmeval(Ahr::HarmonicArray, t; ntrunc, exact = true) -> A::Vector{Matrix}

Evaluate the time response of a harmonic array.

For the harmonic array Ahr representing representing a continuous-time time periodic matrix A(t) and the vector of time values t, A[i] = A(t[i]) is computed for each time value t[i]. If exact = true (default) an exact evaluation is computed, while for exact = false, a linear interpolation based approximation is computed (potentially more accurate in intersample points). The keyword argument ntrunc specifies the number of harmonics to be used for evaluation (default: maximum possible number of harmonics).

 tvmeval(Asym::PeriodicSymbolicMatrix, t) -> A::Vector{Matrix}

Evaluate the time response of a periodic symbolic matrix.

For the periodic symbolic matrix Asym representing a time periodic matrix A(t) and the vector of time values t, A[i] = A(t[i]) is evaluated for each time value t[i].

 tvmeval(Af::PeriodicFunctionMatrix, t) -> A::Vector{Matrix}

Evaluate the time response of a periodic function matrix.

For the periodic function matrix Af representing a time periodic matrix A(t) and the vector of time values t, A[i] = A(t[i]) is evaluated for each time value t[i].

PeriodicSystems.hreval — Function

 hreval(Ahr::HarmonicArray, t; ntrunc, exact = true) -> A::Matrix

Evaluate the harmonic array Ahr representing a continuous-time time periodic matrix A(t) for a numerical or symbolic time value t. For a real value t, if exact = true (default) an exact evaluation is computed, while for exact = false, a linear interpolation based approximation is computed (potentially more accurate in intersample points). The keyword argument ntrunc specifies the number of harmonics to be used for the evaluation (default: maximum possible number). If t is a symbolic variable, a symbolic evaluation of A(t) is performed (see also hr2psm)

PeriodicSystems.pmaverage — Function

pmaverage(A) -> Am

Compute for the continuous-time periodic matrix A(t) the corresponding time averaged matrix Am over one period.