Periodic matrix data types

PeriodicMatrix Discrete-time periodic matrix representation.

PeriodicArray Discrete-time periodic array representation.

PeriodicFunctionMatrix Continuous-time periodic function matrix representation.

PeriodicSymbolicMatrix Continuous-time periodic symbolic matrix representation.

PeriodicTimeSeriesMatrix Continuous-time periodic time series matrix representation.

HarmonicArray Continuous-time harmonic array representation.

PeriodicSystems.PeriodicMatrix — Type

PeriodicMatrix(M, T) -> A::PeriodicMatrix

Discrete-time periodic matrix representation.

The discrete-time periodic matrix object A is built from a p-vector M of real matrices and the associated time period T. M contains the cyclic component matrices M[i], i = 1,..., p, where M[i] represents the value M(Δ(i-1)) of a time periodic matrix M(t) of period T, with Δ := T/p, the associated sampling time. It is assumed that M[k] := M[mod(k-1,p)+1] for arbitrary k. All component matrices are allowed to have arbitrary (time-varying) dimensions. The component matrices M, the period T and the discrete period p can be accessed via A.M, A.period and A.dperiod, respectively.

PeriodicSystems.PeriodicArray — Type

PeriodicArray(M, T) -> A::PeriodicArray

Discrete-time periodic array representation.

The discrete-time periodic array object A is built from a m×n×p real array M and the associated time period T. M contains the cyclic component matrices M[:,:,i], i = 1,..., p, where M[:,:,i] represents the value M(Δ(i-1)) of a time periodic matrix M(t) of period T, with Δ := T/p, the associated sampling time. It is assumed that M[:,:,k] := M[:,:,mod(k-1,p)+1] for arbitrary k. The component matrices M, the period T and the discrete period p can be accessed via A.M, A.period and A.dperiod, respectively.

PeriodicSystems.PeriodicFunctionMatrix — Type

PeriodicFunctionMatrix(f, T) -> A::PeriodicFunctionMatrix

Continuous-time periodic function matrix representation.

The continuous-time periodic function matrix object A is built from a time periodic real matrix function f(t) of real time variable t and the associated time period T. The periodicity condition f(t) ≈ f(t+T) is checked for a randomly generated value of t and a warning is issued if this condition is not satisfied. The function f(t), the period T, and the row and column dimensions of f(t) can be accessed via A.f, A.period and the tuple A.dims, respectively.

PeriodicSystems.PeriodicSymbolicMatrix — Type

PeriodicSymbolicMatrix(F, T) -> A::PeriodicSymbolicMatrix

Continuous-time periodic symbolic matrix representation.

The continuous-time periodic symbolic matrix object A is built from F, a symbolic periodic real matrix or vector of symbolic variable t, and the associated time period T. It is assumed that F(t) = F(t+T) for any real time value t. The symbolic matrix F and the period T can be accessed via A.F and A.period, respectively.

PeriodicSystems.PeriodicTimeSeriesMatrix — Type

PeriodicTimeSeriesMatrix(At, T) -> A::PeriodicTimeSeriesMatrix

Continuous-time periodic time series matrix representation.

The continuous-time periodic time series matrix object A is built from a p-vector At of real matrices and the associated time period T. At contains the cyclic component matrices At[i], i = 1,..., p, where At[i] represents the value A(Δ(i-1)) of a time periodic matrix A(t) of period T, with Δ := T/p, the associated sampling time. It is assumed that At[k] := At[mod(k-1,p)+1] for arbitrary k. All component matrices must have the same dimensions. The component matrices At and the period T can be accessed via A.values and A.period, respectively.

PeriodicSystems.HarmonicArray — Type

 HarmonicArray(Ahr, T) -> A::HarmonicArray

Continuous-time harmonic array representation.

The harmonic array object A is built for the harmonic representation of a periodic matrix Ahr(t) of period T in the form

                 p
 Ahr(t) = A_0 +  ∑ ( Ac_i*cos(i*t*2*π/T)+As_i*sin(i*2*π*t/T) ) .
                i=1

The m×n×(p+1) complex array Ahr contains the harmonic components as follows: Ahr[:,:,1] contains the constant term A_0 (the mean value) and the real and imaginary parts of Ahr[:,:,i+1] for i = 1, ..., p contain the coefficient matrices Ac_i and As_i, respectively. The complex matrix Ahr containing the harmonic components and the period T can be accessed via A.values and A.period, respectively.

PeriodicSystems.HarmonicArray — Method

 HarmonicArray(A0, Ac, As, T) -> A::HarmonicArray

Construct a harmonic array representation from the harmonic components.

The harmonic array object A is built for the harmonic representation Ahr(t) of a periodic matrix of period T in the form

                 p
 Ahr(t) = A_0 +  ∑ ( Ac_i*cos(i*t*2*π/T)+As_i*sin(i*2*π*t/T) ) ,
                i=1

where the constant term A_0 is contained in the real matrix A0, and Ac and As are vectors of real matrices such that the i-th (cosinus) coefficient matrix Ac_i is contained in Ac[i] and the i-th (sinus) coefficient matrix As_i is contained in As[i]. p is the maximum of length of the vectors of matrices Ac and As. If the length of Ac or As is less than p, then zero trailing matrices are assumed in the respective matrix. All component matrices must have the same dimensions. The complex matrix containing the harmonic components and the period T can be accessed via A.values and A.period, respectively.